% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spectral_lowess.R
\name{AIC.graph.spectral.lowess}
\alias{AIC.graph.spectral.lowess}
\title{AIC for Graph Spectral LOWESS}
\usage{
\method{AIC}{graph.spectral.lowess}(object, ..., k = 2)
}
\arguments{
\item{object}{A 'graph.spectral.lowess' object}

\item{...}{Additional arguments (currently unused)}

\item{k}{Penalty parameter (default: 2)}
}
\value{
AIC value
}
\description{
Computes the Akaike Information Criterion
}
