% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spectral_ma_lowess.R
\name{AIC.graph.spectral.ma.lowess}
\alias{AIC.graph.spectral.ma.lowess}
\title{AIC for Graph Spectral MA LOWESS}
\usage{
\method{AIC}{graph.spectral.ma.lowess}(object, ..., k = 2)
}
\arguments{
\item{object}{A 'graph.spectral.ma.lowess' object}

\item{...}{Additional arguments (currently unused)}

\item{k}{Penalty parameter (default: 2)}
}
\value{
AIC value
}
\description{
Computes the Akaike Information Criterion accounting for model averaging
}
