% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{angle.3D}
\alias{angle.3D}
\title{Computes the angle between two 3D vectors in radians.}
\usage{
angle.3D(v, w, method = "arcsine")
}
\arguments{
\item{v}{A 3D vector.}

\item{w}{A 3D vector.}

\item{method}{The method to be used for the angle estimate. Possible choices are: "arcsine" and "arctan".}
}
\value{
The angle between two 3D vectors in radians.
}
\description{
Computes the angle between two 3D vectors in radians.
}
\examples{
\dontrun{
v <- c(0,1,3)
w <- c(2,3,4)
angle <- angle.3D(v, w)
}
}
