% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergences.R
\name{angular.wasserstein.index}
\alias{angular.wasserstein.index}
\title{Compute Angular Wasserstein Index between two point sets}
\usage{
angular.wasserstein.index(X, Y, k)
}
\arguments{
\item{X}{A numeric matrix where each row represents a point in n-dimensional space.}

\item{Y}{A numeric matrix with the same dimensions as X, representing the second point set.}

\item{k}{An integer specifying the number of nearest neighbors to consider.}
}
\value{
A numeric value representing the Angular Wasserstein Index.
}
\description{
This function calculates the Angular Wasserstein Index I_W(X, Y) between two point sets X and Y.
It computes the angular distribution of k-nearest neighbors for each point and then
calculates the Wasserstein distance between these angular distributions.
}
