% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_cx.R
\name{basins.union}
\alias{basins.union}
\title{Union of Gradient Flow Basins}
\usage{
basins.union(x, ids)
}
\arguments{
\item{x}{An object of class "basin_cx" returned by create.basin.cx().}

\item{ids}{Character vector of basin labels (e.g., c("M1", "m2", "M3")).}
}
\value{
An integer vector of vertex indices contained in the union of the specified basins.
}
\description{
Get the union of vertices from specified basins in a basin complex.
}
