% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{bbox.hcube.tiling}
\alias{bbox.hcube.tiling}
\title{Creates a hypercube tiling of a bounding box}
\usage{
bbox.hcube.tiling(w, L, R)
}
\arguments{
\item{w}{A numeric value representing the width of each sub-box, i.e., the edge length.}

\item{L}{A numeric vector representing the left vertices of the bounding box.}

\item{R}{A numeric vector representing the right vertices of the bounding box.}
}
\value{
A list of sub-boxes, where each sub-box is represented as a list with components 'L' and 'R', containing
the left and right vertices of the corresponding sub-box.
}
\description{
This function constructs a hypercube tiling of a bounding box using hypercubes of width 'w'.
}
\details{
The function first ensures that 'L' and 'R' have the same length and then calculates the number of grid elements along
each coordinate. It then creates corresponding intervals and constructs the sub-boxes within the bounding box.
}
\examples{
\dontrun{
w <- 1
L <- c(0, 0)
R <- c(2, 3)
sub_boxes <- bbox.hcube.tiling(w, L, R)
}

}
\seealso{
\code{\link{extract.xy}} for extracting x and y values from a character vector.
}
