% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{bi.gaussian.mixture}
\alias{bi.gaussian.mixture}
\title{Creates a Synthetic Function with Specified Number of Local Maxima using bi-Gaussians.}
\usage{
bi.gaussian.mixture(
  n.lmax,
  x.lmax = NULL,
  y.lmax = NULL,
  x.min = 0,
  x.max = 10,
  y.min = 1,
  y.max = 5,
  min.dist = NULL,
  C.min.dist = 0.5,
  max.itr = 1000,
  n.grid = 400,
  C = 1.5,
  q = 3,
  compact.support = FALSE
)
}
\arguments{
\item{n.lmax}{Number of local maxima.}

\item{x.lmax}{Locations of local maxima.}

\item{y.lmax}{Values of local maxima.}

\item{x.min}{Minimum x value for the grid (defaults to 0).}

\item{x.max}{Maximum x value for the grid (defaults to 10).}

\item{y.min}{Minimum y value for local maxima (defaults to 1).}

\item{y.max}{Maximum y value for local maxima (defaults to 5).}

\item{min.dist}{A positive numeric value specifying the minimum allowable
distance between any two consecutive elements of x.}

\item{C.min.dist}{A real number between 0 and 1 used to specify min.dist if
it is NULL. We use the formula min.dist = C.min.dist * (x.max - x.min) /
n.lmax.}

\item{max.itr}{The maximal number of iterations for finding x.lmin so that
the distance between consecutive elements is no less than min.dist.}

\item{n.grid}{Number of grid points (defaults to 400).}

\item{C}{Scaling factor for the partition of unity (defaults to 2).}

\item{q}{A power parameter of q-Gaussian.}

\item{compact.support}{Set to TRUE to use partition of unity with the components
with compact support and unbound support otherwise.}
}
\value{
A list containing the following components:
\itemize{
\item \code{x}: Vector of x locations for the grid points.
\item \code{y}: Vector of function values at each grid point.
\item \code{x.lmax}: Vector of x locations of the local maxima.
\item \code{y.lmax}: Vector of y values of the local maxima.
}
}
\description{
Generates a synthetic function with specified number of local maxima using bi-Gaussian functions
and a partition of unity. This function is useful for creating complex, non-linear synthetic data
for analysis and testing.
}
\details{
The function generates a set of local maxima points and their corresponding y-values.
It then constructs a partition of unity and a matrix of bi-Gaussian functions centered
at these maxima. The bi-Gaussian function is defined as a Gaussian with sigma.left for
x <= mu and a Gaussian with sigma.right for x > mu. The final synthetic function is
the sum of these bi-Gaussians weighted by the partition of unity.
}
