% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{bi.gaussian}
\alias{bi.gaussian}
\title{Bi-Gaussian defined as Gaussian with sigma.left for x <= mu and a Gaussian
with sigma.right for x > mu}
\usage{
bi.gaussian(x, mu, sigma.left, sigma.right)
}
\arguments{
\item{x}{A vector of x-values.}

\item{mu}{The location of the global maximum of the bi-gaussian function.}

\item{sigma.left}{The standard deviation of the left part of the bi-gaussian function.}

\item{sigma.right}{The standard deviation of the right part of the bi-gaussian function.}
}
\description{
Bi-Gaussian defined as Gaussian with sigma.left for x <= mu and a Gaussian
with sigma.right for x > mu
}
