% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spectral_lowess.R
\name{BIC.graph.spectral.lowess}
\alias{BIC.graph.spectral.lowess}
\title{BIC for Graph Spectral LOWESS}
\usage{
\method{BIC}{graph.spectral.lowess}(object, ...)
}
\arguments{
\item{object}{A 'graph.spectral.lowess' object}

\item{...}{Additional arguments (currently unused)}
}
\value{
BIC value
}
\description{
Computes the Bayesian Information Criterion
}
