% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spectral_ma_lowess.R
\name{BIC.graph.spectral.ma.lowess}
\alias{BIC.graph.spectral.ma.lowess}
\title{BIC for Graph Spectral MA LOWESS}
\usage{
\method{BIC}{graph.spectral.ma.lowess}(object, ...)
}
\arguments{
\item{object}{A 'graph.spectral.ma.lowess' object}

\item{...}{Additional arguments (currently unused)}
}
\value{
BIC value
}
\description{
Computes the Bayesian Information Criterion accounting for model averaging
}
