% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{box.tiling.of.box}
\alias{box.tiling.of.box}
\title{Creates box tiling of a box}
\usage{
box.tiling.of.box(n, box)
}
\arguments{
\item{n}{An integer number specifying the number of sub-intervals (of equal
length within each axis, but of potentially of different lengths between
different axes) of each axis that will be used to construct box tiling of
the given box.}

\item{box}{A box with L, R, i components.}
}
\value{
A list of sub-boxes, where each sub-box is represented as a list with components 'L' and 'R', containing
the left and right vertices of the corresponding sub-box.
}
\description{
This function constructs a box tiling of a box specified by vectors 'L' and 'R'. The boxes forming the tiling are
created by subdividing each axis into the same number of equal length intervals. A box is (also called hyperrectangle
or orthotope) is defined as the Cartesian product of k intervals, where k is the dimension of the orthotope. A box is
defined by two opposite vertices L and R.
}
\examples{
\dontrun{
w <- 1
box <- list()
box$L <- c(0, 0)
box$R <- c(2, 3)
box$i <- 1
sub_boxes <- box.tiling.of.box(w, box)
}

}
\seealso{
\code{\link{extract.xy}} for extracting x and y values from a character vector.
}
