% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{box.tiling}
\alias{box.tiling}
\title{Creates a box tiling of the bounding box of some set.}
\usage{
box.tiling(
  X,
  n.segments.per.axis,
  eps,
  n.itrs = 1,
  p.exp = 0.05,
  plot.it = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{X}{A matrix of points in some Euclidean space.}

\item{n.segments.per.axis}{An integer number specifying the number of
segments (subintervals) of each axis that will be used to construct box
tiling of the given box. The segmenst are of equal length within each
axis, but potentially of different lengths between different axes)}

\item{eps}{A distance threshold (non-negative real number) such that each box
is expanded by eps and then tested for the presence of the elements of X.}

\item{n.itrs}{The number of iterations of box sub-divisions. Default 1.}

\item{p.exp}{An expansion factor for the bounding box. Each edge of the box
is scaled by the factor (1 + p.exp), where p.exp is a non-negative real
number.}

\item{plot.it}{Set to TRUE to see in 2D the progress of the tiling construct.}

\item{verbose}{Set to TRUE for progress messages.}
}
\value{
A list with two components:
\itemize{
\item \code{X.boxes}: A list of boxes covering X, where each box is represented by its left and right vertices.
\item \code{X.vol.ratio}: A numerical vector containing estimates of X's volume ratio over the iterations within this function.
}
}
\description{
This function creates a box tiling of the bounding box of some set by
selecting only boxes of the box tiling containing elements of that set. Box
tiling is an arrangement of boxes without gaps or overlaps, meaning that two
boxes can intersect only along their faces.
}
\examples{
\dontrun{
  X <- matrix(runif(100), ncol = 2)
  boxes <- box.tiling(X, n = 50, eps = 0.05, p.exp = 0.05)
}
}
