% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{chain.with.path}
\alias{chain.with.path}
\title{Create a chain.with.path object}
\usage{
chain.with.path(adj.list, weight.list, gpd.obj)
}
\arguments{
\item{adj.list}{List representing the adjacency structure of the graph}

\item{weight.list}{List of weights corresponding to edges in the adjacency list}

\item{gpd.obj}{Object containing additional graph path data or metadata}
}
\value{
An object of class "chain.with.path" containing the graph structure
and associated data
}
\description{
Constructor function for creating a chain.with.path object that represents
a graph structure with adjacency and weight information, typically used in
path-based algorithms or network analysis.
}
