% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{circle.plot}
\alias{circle.plot}
\title{Create Circle Plot of Matrix Data}
\usage{
circle.plot(
  X,
  n.circle.pts = 100,
  delta = 0.11,
  pch = ".",
  cex = 2.5,
  col = "blue",
  comp.pch = 19,
  comp.col = "red",
  comp.cex = 1.2,
  lab.cex = 1.2,
  edge.col = "gray",
  adj.df = NULL
)
}
\arguments{
\item{X}{A data matrix where rows are observations and columns are variables.}

\item{n.circle.pts}{Number of points used to draw the circle.}

\item{delta}{Margin offset for accommodating labels.}

\item{pch}{Plotting character for data points.}

\item{cex}{Character expansion factor for data points.}

\item{col}{Color of data points.}

\item{comp.pch}{Plotting character for component points on circle.}

\item{comp.col}{Color of component points.}

\item{comp.cex}{Character expansion factor for component points.}

\item{lab.cex}{Character expansion factor for labels.}

\item{edge.col}{Color of edges from center to circle points.}

\item{adj.df}{Matrix of text adjustment values (n.cols x 2) for label positioning.}
}
\value{
Invisibly returns the component points matrix.
}
\description{
Generates a circle plot where matrix columns are mapped to points on a unit circle
}
\details{
This function creates a visualization where the first variable is placed at the
center of a unit circle, and remaining variables are uniformly distributed around
the circle. Data points are then projected into this circular space.
}
\examples{
\dontrun{
X <- matrix(rnorm(50), ncol = 5)
colnames(X) <- paste0("Var", 1:5)
circle.plot(X)
}

}
