% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{clamp}
\alias{clamp}
\title{Restricts the range of value of a numeric vector to \code{[xmin, xmax]}.}
\usage{
clamp(x, xmin, xmax)
}
\arguments{
\item{x}{A numeric vector.}

\item{xmin}{The minimum value of the clamped x.}

\item{xmax}{The maximum value of the clamped x.}
}
\value{
A vector with all values below xmin set to xmin and all value above xmax set to xmax.
}
\description{
Restricts the range of value of a numeric vector to \code{[xmin, xmax]}.
}
