% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fassoc1_test.R
\name{coef.assoc1}
\alias{coef.assoc1}
\title{Coef Method for assoc1 Objects}
\usage{
\method{coef}{assoc1}(object, ...)
}
\arguments{
\item{object}{An object of class "assoc1".}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A named numeric vector containing delta1, Delta1, and p-value.
}
\description{
Extracts key coefficients from an assoc1 object.
}
