% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerve_cx.R
\name{complex.laplacian.solve}
\alias{complex.laplacian.solve}
\title{Solve Full Laplacian Problem on Nerve Complex}
\usage{
complex.laplacian.solve(
  complex,
  lambda = 1,
  dim.weights = rep(1, complex$max_dimension + 1)
)
}
\arguments{
\item{complex}{A nerve complex object}

\item{lambda}{Regularization parameter}

\item{dim.weights}{Weights for each dimension's contribution}
}
\value{
Vector of extended function values
}
\description{
This function solves the full Laplacian problem on the nerve complex to
extend the function values.
}
