% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_smooth_morse_smale.R
\name{compute.gradient.trajectory}
\alias{compute.gradient.trajectory}
\title{Compute Gradient Trajectory}
\usage{
compute.gradient.trajectory(i, j, f.grid)
}
\arguments{
\item{i}{Row index of starting point}

\item{j}{Column index of starting point}

\item{f.grid}{Matrix of function values}
}
\value{
List containing:
\item{ascending.path}{Matrix of (i,j) coordinates along ascending path}
\item{descending.path}{Matrix of (i,j) coordinates along descending path}
\item{local.max}{Coordinates of destination local maximum}
\item{local.min}{Coordinates of destination local minimum}
}
\description{
Computes both ascending and descending gradient trajectories from a grid point.
}
\examples{
\dontrun{
grid <- create.grid(30)
f <- function(x, y) -((x-0.5)^2 + (y-0.5)^2)
f.grid <- evaluate.function.on.grid(f, grid)
traj <- compute.gradient.trajectory(15, 15, f.grid)
str(traj)
}
}
