% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_smooth_morse_smale.R
\name{compute.morse.smale.cells}
\alias{compute.morse.smale.cells}
\title{Compute Morse-Smale Cells}
\usage{
compute.morse.smale.cells(f.grid)
}
\arguments{
\item{f.grid}{Matrix of function values}
}
\value{
Matrix of cell labels
}
\description{
Assigns each grid point to a Morse-Smale cell based on gradient flow.
}
\examples{
\dontrun{
grid <- create.grid(30)
f <- function(x, y) sin(3*x) * cos(3*y)
f.grid <- evaluate.function.on.grid(f, grid)
cells <- compute.morse.smale.cells(f.grid)
image(cells)
}
}
