% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_cx.R
\name{compute_asc_desc_cell_intersection_matrix}
\alias{compute_asc_desc_cell_intersection_matrix}
\title{Compute Ascending-Descending Cell Intersection Matrix}
\usage{
compute_asc_desc_cell_intersection_matrix(basin_cx)
}
\arguments{
\item{basin_cx}{An object of class 'basin_cx'}
}
\value{
A square matrix where (i,j) is the number of shared vertices between cell i and cell j,
with labels as lmin_label:lmax_label. Final row and column contain cell sizes.
}
\description{
Compute Ascending-Descending Cell Intersection Matrix
}
