% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{convert.adjacency.list.to.adjacency.matrix}
\alias{convert.adjacency.list.to.adjacency.matrix}
\title{Converts a graph adjacency list to an adjacency matrix}
\usage{
convert.adjacency.list.to.adjacency.matrix(adj.list, weights.list = NULL)
}
\arguments{
\item{adj.list}{A list representing the adjacency list of a graph. Each element of the list
is a vector of node indices that the corresponding node is connected to.}

\item{weights.list}{(optional) A list of the same length as \code{adj.list}, where each element
is a vector of weights for the edges connecting the corresponding node to the nodes in \code{adj.list}.
If not provided, the adjacency matrix will be binary (1 for edges and 0 for no edges).}
}
\value{
A matrix representing the adjacency matrix of the graph.
}
\description{
This function converts a given graph adjacency list into an adjacency matrix,
where each entry in the matrix indicates whether there is an edge between the corresponding nodes.
}
\examples{
adj.list <- list(c(2, 4), c(1, 3), c(2, 4), c(1, 3))
weights.list <- list(c(2, 4), c(2, 3), c(3, 1), c(4, 1))
convert.adjacency.list.to.adjacency.matrix(adj.list)
convert.adjacency.list.to.adjacency.matrix(adj.list, weights.list)
}
