% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{convert.adjacency.matrix.to.adjacency.list}
\alias{convert.adjacency.matrix.to.adjacency.list}
\title{Converts a graph adjacency matrix to an adjacency list}
\usage{
convert.adjacency.matrix.to.adjacency.list(M)
}
\arguments{
\item{M}{A matrix representing the adjacency matrix of a graph.}
}
\value{
A list representing the adjacency list of the graph.
}
\description{
This function converts a given graph adjacency matrix into an adjacency list,
where each entry in the list corresponds to a node and contains the indices
of nodes it is connected to.
}
