% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{convert.adjacency.to.edgelist}
\alias{convert.adjacency.to.edgelist}
\title{Convert an Adjacency List to an Edgelist}
\usage{
convert.adjacency.to.edgelist(adj.list)
}
\arguments{
\item{adj.list}{An adjacency list where each element is a vector of indices
representing the neighbors of a node.}
}
\value{
A list where each element is a vector of length two, representing
an edge in the graph (source node, target node).
}
\description{
This function takes an adjacency list representation of a graph and converts it
into an edgelist format suitable for use with plotting functions like \code{igraph}.
}
