% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{convert.edge.label.list.to.edge.label.vector}
\alias{convert.edge.label.list.to.edge.label.vector}
\title{Converts an Edge Label List to an Edge Vector}
\usage{
convert.edge.label.list.to.edge.label.vector(
  edge.label.list,
  rm.duplicates = TRUE
)
}
\arguments{
\item{edge.label.list}{A list of edge labels.}

\item{rm.duplicates}{Set to TRUE to allow for duplicate edges.}
}
\value{
A vector of edge labels.
}
\description{
This function takes a list of edge labels and turns it into a vector.
}
