% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{convert.to.undirected}
\alias{convert.to.undirected}
\title{Converts a directed graph to an undirected graph}
\usage{
convert.to.undirected(adj.list)
}
\arguments{
\item{adj.list}{A named list representing the adjacency list of the directed graph.
Each element of the list is a character vector containing the names
of the adjacent vertices for a given vertex.}
}
\value{
A named list representing the adjacency list of the corresponding undirected graph.
Each element of the list is a character vector containing the names of the adjacent
vertices for a given vertex in the undirected graph.
}
\description{
This function takes the adjacency list representation of a directed graph and
converts it into the adjacency list representation of the corresponding undirected graph.
}
\examples{
directed_graph <- list(
  "A" = c("B", "C"),
  "B" = c("C"),
  "C" = c("D"),
  "D" = c("A")
)

undirected_graph <- convert.to.undirected(directed_graph)
print(undirected_graph)

}
