% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{convert.weighted.adjacency.matrix.to.adjacency.list}
\alias{convert.weighted.adjacency.matrix.to.adjacency.list}
\title{Converts a graph weighted adjacency matrix to an adjacency list and weights list}
\usage{
convert.weighted.adjacency.matrix.to.adjacency.list(M)
}
\arguments{
\item{M}{A matrix representing the weighted adjacency matrix of a graph.}
}
\value{
A list with two elements: \code{adjacency.list} and \code{weights.list}.
\code{adjacency.list} is a list where each element is a vector of connected node indices.
\code{weights.list} is a list where each element is a vector of weights corresponding to the edges in the adjacency list.
}
\description{
This function converts a given weighted graph adjacency matrix into two lists:
\enumerate{
\item An adjacency list where each entry corresponds to a node and contains the indices
of nodes it is connected to.
\item A weights list where each entry corresponds to a node and contains the weights
of the edges connecting it to other nodes.
}
}
