% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{create.2D.grid}
\alias{create.2D.grid}
\title{Creates a 2D grid of points around X.}
\usage{
create.2D.grid(n, X, f = 0.2, eSDf = 1.5, gRf = 4)
}
\arguments{
\item{n}{The number of uniformly spaced points, seq(min(xi), max(xi), length=n),
on each axis that are the basis of the grid.}

\item{X}{A set of points around which the grid is created if not NULL.}

\item{f}{A fraction of x1 and x2 range that the grid is extended to.}

\item{eSDf}{An edge subdivision factor that is a scaling factor, such that edges of mstree(X) are
subdivided if their length is greater than eSDf * mode(edge.len).}

\item{gRf}{A grid radius factor that is a scaling factor, such that points of the rectangular grid
that are further than gRf*mode(edge.len) away from the closest
subdivision of the minimal spanning tree of X, mstree(X), are eliminated from the grid.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame, grid, with n*n rows and 2 columns of the grid points in
        2D. If X is not NULL, grid contains only points nor more than eps away
        from the mstree(X).
}\if{html}{\out{</div>}}
}
\description{
If X is NULL, x1.range and x2.range will be used to create a rectangular grid of n*n points.
}
