% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{create.3D.grid}
\alias{create.3D.grid}
\title{Creates a 3D grid of points around X}
\usage{
create.3D.grid(n, X, f = 0.2, min.gSf = 1.5, eSDf = 1.5, gRf = 2)
}
\arguments{
\item{n}{The number of uniformly spaced points, seq(min(xi), max(xi), length=n),
on each axis that are the basis of the grid.}

\item{X}{A set of points around which the grid is created if not NULL.}

\item{f}{A fraction of x1, x2 and x3 range that the grid is extended to.}

\item{min.gSf}{A scaling factor, indicating the minimal size of
the grid around X, with respect to the size of X. That is the minimal
size of the grid needs to be at least min.gSf*|X|, where |X|
is the size of X. For example, with min.gSf set to 1.5, the
mimal size of the grid will be, if possible, at least 1.5*|X|. For this
to be possible the initial rectangular grid size has to be at least of
that minimal size, but in practice much bigger. It overrides
gRf if the size of the grid given by the value of
gRf is not}

\item{eSDf}{An edge subdivisioin factor that is a scaling factor, such that edges of mstree(X) are
subdivided is their length is greater than eSDf * mode(edge.len).}

\item{gRf}{A scaling factor, such that points of the rectangular grid
that are further than gRf*mode.edge.len away from the closest
subdivision of the minimal spanning tree of X, mstree(X), are eliminated from the grid.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A data frame, grid, with n*n*n rows and 3 columns of the grid points in
               3D. If X is not NULL, grid contains only points nor more than eps away
               from the mstree(X).
}\if{html}{\out{</div>}}
}
\description{
Creates a 3D grid of points around X
}
