% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{create.3D.grid.v2}
\alias{create.3D.grid.v2}
\title{Creates a 3D grid of points around X.}
\usage{
create.3D.grid.v2(
  n,
  X,
  f = 0.2,
  mst.grid = NULL,
  mode.edge.len = NULL,
  min.gSf = 1.5,
  min.min.gSf = 1,
  gRf = 2,
  min.gRf = 1,
  eSDf = 1.5,
  max.n = 120,
  verbose = FALSE
)
}
\arguments{
\item{n}{The number of uniformly spaced points, seq(min(xi), max(xi), length=n),
on each axis that are the basis of the grid.}

\item{X}{A set of points around which the grid is created if not NULL.}

\item{f}{A fraction of x1, x2 and x3 range that the grid is extended to.}

\item{mst.grid}{A matrix of 3D points that are subdivisions of edges of the mstree(X).}

\item{mode.edge.len}{The mode of the edge lengths.}

\item{min.gSf}{A scaling factor, indicating the minimal size of
the grid around X, with respect to the size of X. That is the minimal
size of the grid needs to be at least min.gSf*|X|, where |X|
is the size of X. For example, with min.gSf set to 1.5, the
minimal size of the grid will be, if possible, at least 1.5*|X|. For this
to be possible the initial rectangular grid size has to be at least of
that minimal size, but in practice much bigger. It overrides
gRf if the size of the grid given by the value of
gRf is not}

\item{min.min.gSf}{The minimal allowable value of min.gSf.}

\item{gRf}{A scaling factor, such that points of the rectangular grid
that are further than gRf*mode.edge.len away from the closest
subdivision of the minimal spanning tree of X, mstree(X), are eliminated from the grid.}

\item{min.gRf}{The minimal allowable value of gRf.}

\item{eSDf}{A scaling factor, such that edges of mstree(X) are
subdivided is their length is greater than eSDf * mode(edge.len).}

\item{max.n}{The maximal value of n that can be reached as n is adjusted to
satisfy the size and radius factor conditions.}

\item{verbose}{Set to TRUE to see info on what is being done.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A data frame, grid, with n*n*n rows and 3 columns of the grid points in
               3D. If X is not NULL, grid contains only points nor more than eps away
               from the mstree(X).
}\if{html}{\out{</div>}}
}
\description{
In this version if the min.gSf condition is not satisfied, n is
increases until it satisfies that condition. Thus, gRf is never modified.
}
