% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{create.3D.rect.grid}
\alias{create.3D.rect.grid}
\title{Creates a rectangular 3D grid}
\usage{
create.3D.rect.grid(
  n,
  x1.range,
  x2.range,
  x3.range,
  type = c("unif", "runif", "norm"),
  f = 0.2
)
}
\arguments{
\item{n}{The number of uniformly spaced points, seq(min(xi), max(xi), length=n),
on each axis that are the basis of the grid.}

\item{x1.range}{A range of x1 values - the first coordinate.}

\item{x2.range}{A range of x2 values - the second coordinate.}

\item{x3.range}{A range of x2 values - the second coordinate.}

\item{type}{A type of distribution the points are sampled from.}

\item{f}{A fraction of x1 and x2 range that the grid is extended to.}
}
\description{
Creates a rectangular 3D grid
}
