% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{create.3D.TN.grid.plus}
\alias{create.3D.TN.grid.plus}
\title{Creates a tubular neighborhood 3D grid of X returning a list with a grid
around X and the distance to the boundary of the grid.}
\usage{
create.3D.TN.grid.plus(
  X,
  mst.grid,
  dx,
  mode.edge.len,
  gRf,
  f = 0.05,
  verbose = FALSE
)
}
\arguments{
\item{X}{A set of points around which the grid is created if not NULL.}

\item{mst.grid}{A matrix of 3D points that are subdivisions of edges of the mstree(X).}

\item{dx}{The length of the grid edge.}

\item{mode.edge.len}{The mode of the edge lengths.#'}

\item{gRf}{A scaling factor, such that points of the rectangular grid
that are further than gRf*mode.edge.len away from the closest
subdivision of the minimal spanning tree of X, mstree(X), are eliminated from the grid.}

\item{f}{A fraction of x1, x2 and x3 range that the grid is extended to.}

\item{verbose}{Set to TRUE to see info on what is being done.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    A data frame, grid, with n*n*n rows and 3 columns of the grid points in
               3D. If X is not NULL, grid contains only points nor more than eps away
               from the mstree(X).
}\if{html}{\out{</div>}}
}
\description{
Creates a tubular neighborhood 3D grid of X returning a list with a grid
around X and the distance to the boundary of the grid.
}
