% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{create.adaptive.tiled.X.grid.xD}
\alias{create.adaptive.tiled.X.grid.xD}
\title{Create an Adaptive Tiled Grid Representation of X}
\usage{
create.adaptive.tiled.X.grid.xD(
  X,
  gSf,
  gRf,
  n = 5,
  n.itrs = 1,
  p.exp = 0.075,
  K = 10,
  wC = 0.1,
  wF = 1.25,
  X.vol.frac = NULL,
  vol.box.grid.size = 10^3,
  max.box.grid.size = 10^7,
  min.tree.obj = NULL,
  verbose = FALSE
)
}
\arguments{
\item{X}{A numeric matrix where each row represents a point in the space.}

\item{gSf}{Grid size factor. Determines the target number of grid points relative to the number of input points.}

\item{gRf}{Grid radius factor. Controls the proximity of grid points to the minimal spanning tree of X.}

\item{n}{Number of boxes for the initial tiling. Default is 5.}

\item{n.itrs}{Number of iterations for the box tiling process. Default is 1.}

\item{p.exp}{Expansion factor for the bounding box. Must be between 0 and 1. Default is 0.075.}

\item{K}{Number of nearest neighbors to consider in the box tiling process. Default is 10.}

\item{wC}{Multiplication factor for temporary grid creation. Must be positive. Default is 0.1.}

\item{wF}{Factor for increasing grid width if necessary. Default is 1.25.}

\item{X.vol.frac}{Pre-computed volume fraction of X, if available. Default is NULL.}

\item{vol.box.grid.size}{Target size for the grid used in volume estimation. Default is 1000.}

\item{max.box.grid.size}{Maximum allowed size for a single box grid. Default is 10^7.}

\item{min.tree.obj}{Pre-computed minimal spanning tree object, if available. Default is NULL.}

\item{verbose}{Logical; if TRUE, print progress information. Default is FALSE.}
}
\value{
A list of class "gridX" containing:
\item{X.grid}{The generated grid points}
\item{d.grid}{Distance values for the grid points}
\item{mst.grid}{Minimal spanning tree of X}
\item{mode.edge.len}{Mode of the edge lengths in the minimal spanning tree}
\item{L, R}{Lower and upper bounds of the bounding box}
\item{dim}{Dimension of the space}
\item{X}{Original input data}
\item{X.vol.frac}{Estimated volume fraction of X}
\item{X.vol}{Estimated volume of X}
\item{gRf, gSf, p.exp, wC}{Input parameters}
}
\description{
This function creates a grid representation of the input data X using an adaptive
tiling approach. It's particularly effective for large or complex datasets.
}
\examples{
\dontrun{
X <- matrix(rnorm(1000), ncol = 2)
grid_obj <- create.adaptive.tiled.X.grid.xD(X, gSf = 2, gRf = 1.5, verbose = TRUE)
plot(grid_obj$X.grid, col = "red", pch = 20)
points(X, col = "blue", pch = 1)
}

}
\seealso{
\code{\link{create.X.grid.xD}} for the original, non-tiled version
}
