% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_generators.R
\name{create.bi.kNN.chain.graph}
\alias{create.bi.kNN.chain.graph}
\title{Creates a bi-k-NN chain graph}
\usage{
create.bi.kNN.chain.graph(n.vertices = 5, k = 1, x = NULL, y = NULL)
}
\arguments{
\item{n.vertices}{An integer. The number of vertices in the graph.}

\item{k}{An integer. The number of neighbors to connect on each side of a vertex.}

\item{x}{An optional numeric vector of length n.vertices. If provided, vertices
will be ordered based on these values.}

\item{y}{An optional numeric vector of length n.vertices. If provided along with x,
it will be sorted according to x.}
}
\value{
A list containing:
\item{adj.list}{A list representing the graph structure. Each element is a vector
of indices of neighboring vertices.}
\item{edge.lengths}{A list of edge lengths corresponding to the graph structure.
If x is provided, these are based on differences in x values; otherwise, they
are set to 1.}
\item{x.sorted}{The sorted x vector if x was provided, otherwise NULL.}
\item{y.sorted}{The y vector sorted according to x if both x and y were provided,
otherwise NULL.}
}
\description{
This function constructs a bi-directional k-nearest neighbor chain graph.
Each vertex is connected to up to k neighbors on both sides, forming a
chain-like structure. The graph can be based on sorted x-coordinates if provided.
}
\examples{
# Create a 5-chain graph with 10 vertices
graph <- create.bi.kNN.chain.graph(10, 5)

# Create a graph based on x-coordinates
x <- runif(10)
y <- sin(x)
graph.with.coords <- create.bi.kNN.chain.graph(10, 2, x, y)

}
