% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_generators.R
\name{create.bipartite.graph}
\alias{create.bipartite.graph}
\title{Create a Bipartite Graph}
\usage{
create.bipartite.graph(n1, n2)
}
\arguments{
\item{n1}{An integer specifying the number of nodes in the first set.}

\item{n2}{An integer specifying the number of nodes in the second set.}
}
\value{
A list representing the adjacency list of the bipartite graph.
The first n1 elements correspond to nodes in the first set,
and the next n2 elements correspond to nodes in the second set.
}
\description{
Generates a bipartite graph with two disjoint sets of nodes.
}
\details{
In a bipartite graph, nodes are divided into two disjoint sets,
and each edge connects a node in the first set to a node in the second set.
There are no edges between nodes in the same set.
}
\examples{
bipartite_graph <- create.bipartite.graph(3, 4)

}
