% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_generators.R
\name{create.chain.graph.with.offset}
\alias{create.chain.graph.with.offset}
\title{Creates a chain graph with n vertices}
\usage{
create.chain.graph.with.offset(n, offset = 0)
}
\arguments{
\item{n}{The number of vertices in the graph}

\item{offset}{An offset in indexing the vertices of the graph.}
}
\value{
A chain graph.
}
\description{
Creates a chain graph with n vertices
}
