% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_generators.R
\name{create.circular.graph}
\alias{create.circular.graph}
\title{Creates a circular graph with n vertices}
\usage{
create.circular.graph(n)
}
\arguments{
\item{n}{The number of vertices in the graph}
}
\value{
A circular graph.
}
\description{
Creates a circular graph with n vertices
}
