% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_generators.R
\name{create.complete.graph}
\alias{create.complete.graph}
\title{Create a Complete Graph}
\usage{
create.complete.graph(n)
}
\arguments{
\item{n}{An integer specifying the number of nodes in the graph.}
}
\value{
A list representing the adjacency list of the complete graph.
Each element of the list corresponds to a node and contains
a vector of all other nodes it's connected to.
}
\description{
Generates a complete graph where every node is connected to every other node.
}
\details{
In a complete graph of n nodes, each node has n-1 connections (to all other nodes).
The resulting graph is undirected.
}
\examples{
complete_graph <- create.complete.graph(5)

}
