% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{create.ED.grid.2D}
\alias{create.ED.grid.2D}
\title{Creates a equi-distant (all edges are of equal length) 2D grid using a C routine}
\usage{
create.ED.grid.2D(dx, x1.range, x2.range, f = 0.2, verbose = FALSE)
}
\arguments{
\item{dx}{The length of each edge.}

\item{x1.range}{A range of x1 values - the first coordinate.}

\item{x2.range}{A range of x2 values - the second coordinate.}

\item{f}{A fraction of x1 and x2 range that the grid is extended to.}

\item{verbose}{If TRUE, some debugging messages will be printed.}
}
\description{
Creates a equi-distant (all edges are of equal length) 2D grid using a C routine
}
