% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{create.ED.grid.xD}
\alias{create.ED.grid.xD}
\title{Create an Equidistant Grid in x-Dimensional Bounding Box (C Interface)}
\usage{
create.ED.grid.xD(edge.length, lower.bounds, upper.bounds)
}
\arguments{
\item{edge.length}{A positive numeric value representing the common length of each edge in the grid.}

\item{lower.bounds}{A numeric vector containing the left (lower) end of the range for each dimension.}

\item{upper.bounds}{A numeric vector containing the right (upper) end of the range for each dimension.}
}
\value{
A matrix representing the equidistant grid, where each row corresponds to a point in the grid,
and each column corresponds to a dimension.
}
\description{
This function generates an equidistant grid within a specified x-dimensional bounding box.
The grid consists of points that are uniformly spaced, and each edge of the grid has the same length.
}
\examples{
\dontrun{
edge.length <- 1.0
lower.bounds <- c(0, 0, 0)
upper.bounds <- c(2, 3, 4)
grid <- create.ED.grid.xD(edge.length, lower.bounds, upper.bounds, round.up = TRUE)
print(grid)
}
}
