% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_generators.R
\name{create.empty.graph}
\alias{create.empty.graph}
\title{Create an Empty Graph}
\usage{
create.empty.graph(n)
}
\arguments{
\item{n}{An integer specifying the number of nodes in the graph.}
}
\value{
A list representing the adjacency list of the empty graph.
Each element of the list is an empty vector, indicating no connections.
}
\description{
Generates an empty graph with a specified number of nodes but no edges.
}
\details{
An empty graph contains nodes but no edges connecting these nodes.
This function is useful for initializing graphs or testing edge cases in graph algorithms.
}
\examples{
empty_graph <- create.empty.graph(5)

}
