% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{create.ENPs.grid.2D}
\alias{create.ENPs.grid.2D}
\title{Creates a equi-number (within each axis) 2D grid using a C routine}
\usage{
create.ENPs.grid.2D(n, x1.range, x2.range, f = 0.2)
}
\arguments{
\item{n}{The number of uniformly spaced points, seq(min(xi), max(xi), length=n),
on each axis that are the basis of the grid.}

\item{x1.range}{A range of x1 values - the first coordinate.}

\item{x2.range}{A range of x2 values - the second coordinate.}

\item{f}{A fraction of x1 and x2 range that the grid is extended to.}
}
\description{
Creates a equi-number (within each axis) 2D grid using a C routine
}
