% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{create.final.grid}
\alias{create.final.grid}
\title{Creates a Final Grid within Selected Boxes}
\usage{
create.final.grid(boxes, X, w, epsilon)
}
\arguments{
\item{boxes}{A list of selected sub-boxes, each represented as a list containing vectors \code{L} and \code{R} for left and right boundaries.}

\item{X}{A matrix representing the original dataset for which the grid is to be created.}

\item{w}{A numeric value representing the width of the grid.}

\item{epsilon}{A numeric value representing the maximum distance from \code{X} that a grid point can be to be included in the final grid.}
}
\value{
A matrix representing the final filtered grid.
}
\description{
This function creates a uniform grid within a set of selected sub-boxes, then filters the grid to remove points that are farther than a given distance \code{epsilon} from the original dataset \code{X}, and removes duplicated grid elements.
}
\examples{
\dontrun{
boxes <- create.ED.boxes(0.5, c(0,0), c(1,1))
X <- matrix(runif(20), ncol = 2)
w <- 0.1
epsilon <- 0.5
final_grid <- create.final.grid(boxes, X, w, epsilon)
}
}
