% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_smooth_morse_smale.R
\name{create.morse.smale.complex}
\alias{create.morse.smale.complex}
\title{Create Morse-Smale Complex Object}
\usage{
create.morse.smale.complex(f.grid)
}
\arguments{
\item{f.grid}{Matrix of function values}
}
\value{
S3 object of class "morse.smale.complex" with components:
\item{cell_lookup}{List mapping cell IDs to point coordinates}
\item{cell_summary}{Data frame summarizing each cell}
\item{local_maxima}{Matrix of local maximum coordinates}
\item{local_minima}{Matrix of local minimum coordinates}
\item{find_cell}{Function to find cell ID for a grid point}
\item{get_neighboring_cells}{Function to find neighboring cells}
}
\description{
Constructs complete Morse-Smale complex from grid function values.
}
\examples{
\dontrun{
grid <- create.grid(30)
f <- function(x, y) sin(3*x) * cos(3*y)
f.grid <- evaluate.function.on.grid(f, grid)
complex <- create.morse.smale.complex(f.grid)
str(complex)
}
}
