% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerve_cx.R
\name{create.nerve.complex}
\alias{create.nerve.complex}
\title{Creates a Nerve Complex Associated With kNN Covering of a Dataset}
\usage{
create.nerve.complex(X, k, max.dim = 2)
}
\arguments{
\item{X}{Matrix of point coordinates, where rows are points and columns are dimensions}

\item{k}{Number of nearest neighbors to use for the covering}

\item{max.dim}{Maximum dimension of simplices to compute}
}
\value{
A nerve complex object
}
\description{
This function creates a nerve complex from a set of points in \eqn{R^n} based on their
k-nearest neighbor covering.
}
