% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{create.random.graph}
\alias{create.random.graph}
\title{Create a Random Graph}
\usage{
create.random.graph(n_vertices, avg_degree, connected = TRUE)
}
\arguments{
\item{n_vertices}{Number of vertices in the graph}

\item{avg_degree}{Average number of neighbors per vertex}

\item{connected}{Logical; if TRUE, ensure the graph is connected}
}
\value{
A list with adjacency and weight lists
}
\description{
Creates a random graph with the specified number of vertices and
average degree (neighbors per vertex).
}
\examples{
graph <- create.random.graph(100, 4)

}
