% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_generators.R
\name{create.star.graph}
\alias{create.star.graph}
\title{Create a star graph by joining chain graphs}
\usage{
create.star.graph(sizes)
}
\arguments{
\item{sizes}{A vector of positive integers specifying the sizes of the chain graphs to be
joined to the central vertex. Each size represents the number of vertices in
a chain graph.}
}
\value{
A list representing the adjacency list of the created star graph.
}
\description{
This function creates a star graph by joining chain graphs of specified sizes to a central
vertex. The resulting star graph consists of a central vertex connected to multiple chains
of vertices.
}
\examples{
\dontrun{
star_graph <- create.star.graph(c(3, 4, 2))
print(star_graph)
}
}
