% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{create.subgraph}
\alias{create.subgraph}
\title{Create a subgraph from a given graph}
\usage{
create.subgraph(
  S.graph,
  id.indices = NULL,
  ids = NULL,
  S = NULL,
  use.sequential.indices = FALSE
)
}
\arguments{
\item{S.graph}{A list containing the original graph structure with adjacency and distance lists.}

\item{id.indices}{A vector of indices to include in the subgraph. Default is NULL.}

\item{ids}{A vector of IDs to include in the subgraph. Default is NULL.}

\item{S}{A data frame or matrix where rownames correspond to node IDs. Required if \code{ids} is provided. Default is NULL.}

\item{use.sequential.indices}{Logical. If TRUE, renumber the indices in the subgraph to be 1:length(id.indices). Default is FALSE.}
}
\value{
A list containing the subgraph structure with adjacency and distance lists.
}
\description{
This function creates a subgraph from a given graph based on specified indices or IDs.
}
\details{
The function can create a subgraph based on either \code{id.indices} or \code{ids}. If \code{ids} is provided,
\code{S} must also be provided to map the IDs to indices. The \code{use.sequential.indices} parameter
allows for renumbering the indices in the subgraph to be sequential, which can be useful for
certain applications.
}
\examples{
X <- runif.sphere(20, 2)
graph <- create.single.iknn.graph(X, k = 3, compute.full = TRUE, verbose = FALSE)
graph$dist_list <- graph$weight_list
graph$weight_list <- NULL
subgraph <- create.subgraph(graph, id.indices = c(1:10))
subgraph_sequential <- create.subgraph(graph, id.indices = c(1:10, 15:16),
                                       use.sequential.indices = TRUE)
}
