% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesics.R
\name{create.tubular.nbhd.of.a.geodesic}
\alias{create.tubular.nbhd.of.a.geodesic}
\title{Creates a tubular neighborhood of a geodesic}
\usage{
create.tubular.nbhd.of.a.geodesic(
  smoothed.geodesic,
  S,
  d.thld = "auto",
  dist.to.geodesic.q = 0.25
)
}
\arguments{
\item{smoothed.geodesic}{A matrix of smoothed geodesic points in S.}

\item{S}{A 3d model of a state space.}

\item{d.thld}{A distance to the geodesic threshold. Set to 'auto' for automatic determination of the threshold.}

\item{dist.to.geodesic.q}{The quantile of the distances to the geodesic to which the density function of the distances will be restricted to.}
}
\value{
A list containing:
\item{nn.i}{A vector of indices indicating the nearest point on the geodesic for each point in S.}
\item{dist.to.geodesic}{A vector of distances from each point in S to the geodesic.}
\item{strand.i}{Indices of points in S that are within the tubular neighborhood.}
\item{strand}{The subset of S containing only points within the tubular neighborhood.}
\item{nn.geodesic.vertex.i}{Indices of the nearest geodesic vertices for points in the strand.}
\item{dist.along.strand.geodesic}{Normalized distances along the geodesic for points in the strand.}
\item{d.thld}{The distance threshold used to define the tubular neighborhood.}
}
\description{
Creates a tubular neighborhood of a geodesic
}
\details{
This function creates a tubular neighborhood around a geodesic path by identifying
all points in the state space S that are within a specified distance threshold of
the geodesic. When d.thld is set to "auto", the threshold is determined based on
the distribution of distances to the geodesic.
}
