% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{create.X.grid}
\alias{create.X.grid}
\title{Creates a grid around a state space.}
\usage{
create.X.grid(X, gSf, gRf, min.K = 10, med.dK.divf = 5, max.dx.C = 1)
}
\arguments{
\item{X}{A state space for which the grid is to be created.}

\item{gSf}{A grid size scaling factor.}

\item{gRf}{A grid radius scaling factor. Points on the rectangular grid that
are farther than gRf*mode.edge.len from the closest subdivision of
the minimal spanning tree of X, mstree(X), are eliminated from the grid.}

\item{min.K}{The minimum number of nearest neighbors ('x' NNs) that must be present in each window.}

\item{med.dK.divf}{A division factor for the lower bound of bandwidth (bw) and dx. It is equal to
med.dK/med.dK.divf, where \code{med.dK=median(dK)} and \code{dK=nn.dist[,min.K]}.}

\item{max.dx.C}{A division factor for estimating max.dx, defined as the length of the diagonal in
the X enclosing box divided by max.dx.C.}
}
\value{
A list with the following components:
\itemize{
\item \code{X.grid}: A uniform grid in a tubular neighborhood of X.
\item \code{d.grid}: The distance to the boundary of the grid.
\item \code{mst.grid}: A uniform grid over the edges of the minimum spanning tree of X.
\item \code{opt.dx}: The optimal value of the dx parameter for given gSf and gRf values.
\item \code{mode.edge.len}: The mode of the lengths of the edges of the minimum spanning tree of X.
}
}
\description{
This function generates a uniform grid in a tubular neighborhood of a state space.
}
\examples{
\dontrun{
# Let X be a low-dimensional model of a state space.
res <- create.X.grid(X, gSf=5, gRf=5, min.K=10, med.dK.divf=5, max.dx.C=1)
str(res)
}
}
