% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{create.X.grid.xD}
\alias{create.X.grid.xD}
\title{Creates a Uniform Grid in a Tubular Neighborhood of a State Space}
\usage{
create.X.grid.xD(
  X,
  gSf,
  gRf,
  p.exp = 0.05,
  wC = 0.1,
  wF = 1.25,
  X.vol.frac = NULL,
  max.tmp.grid.size = 10^7,
  min.tree.obj = NULL,
  n.threads = 1,
  verbose = FALSE
)
}
\arguments{
\item{X}{A numeric matrix representing the state space for which the grid is to be created.}

\item{gSf}{A numeric value representing the grid size scaling factor, such that the grid size \eqn{n(X.grid) = gSf * n(X)}, where \eqn{n(X)} is the number of rows of X.}

\item{gRf}{A numeric value representing the grid radius scaling factor. Points on the rectangular grid that are farther than \eqn{gRf * mode.edge.len} from the closest subdivision of the minimal spanning tree of X, \eqn{mstree(X)}, are eliminated from the grid.}

\item{p.exp}{A numeric value representing the expansion factor of the bounding box. Must be between 0 and 1. Default is 0.05.}

\item{wC}{A numeric value representing a multiplication factor controlling the precision of the estimation of the volume of X. Must be a positive number. Default is 0.1.}

\item{wF}{A numeric value representing a weighting factor. Default is 1.25.}

\item{X.vol.frac}{The volume fraction of X in its bounding box.}

\item{max.tmp.grid.size}{Maximum size of temporary grid. Default is 10^7.}

\item{min.tree.obj}{A minimal spanning tree object. This parameter is useful when running this function with different values of gSf and gRf parameters and so only the first call may have min.tree.obj set to NULL and all remaining may use a minimal tree object from the first call.}

\item{n.threads}{The number of threads to use in the corresponding C function call.}

\item{verbose}{A logical value. Set to TRUE to see information on what is being done. Default is FALSE.}
}
\value{
A list with class \code{"gridX"} and the following components:
\itemize{
\item \code{X.grid}: A numeric matrix representing a uniform grid in a tubular neighborhood of X.
\item \code{d.grid}: A numeric vector representing the distance to the boundary of the grid.
\item \code{mst.grid}: A numeric matrix representing a uniform grid over the edges of the minimum spanning tree of X.
\item \code{mode.edge.len}: A numeric value representing the mode of the lengths of the edges of the minimum spanning tree of X.
\item \code{L}: A numeric vector representing the left end of the range for each dimension.
\item \code{R}: A numeric vector representing the right end of the range for each dimension.
\item \code{dim}: An integer representing the number of dimensions.
\item \code{X}: A numeric matrix representing the original state space.
\item \code{X.vol.frac}: A numeric value representing the fraction of the volume of X in the bounding box.
\item \code{X.vol}: A numeric value representing the estimated volume of X.
\item \code{gRf}: A numeric value representing the grid radius scaling factor.
\item \code{gSf}: A numeric value representing the grid size scaling factor.
\item \code{p.exp}: A numeric value representing the expansion factor of the bounding box.
\item \code{wC}: A numeric value representing the multiplication factor controlling the precision of the volume estimation.
}
}
\description{
This function generates a uniform grid in a tubular neighborhood of a given state space X, considering specified parameters for grid size, radius scaling, and other factors.
}
\examples{
\dontrun{
# Let X be a low-dimensional model of a state space.
res <- create.X.grid.xD(X, gSf=5, gRf=5, p.exp=0.05, wC=0.1)
str(res)
}
}
