% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{cross.prod}
\alias{cross.prod}
\title{Computes a cross product between two 3D vectors}
\usage{
cross.prod(x, y)
}
\arguments{
\item{x}{A 3D vector.}

\item{y}{A 3D vector.}
}
\value{
The cross product of x and y.
}
\description{
This routine calculates the cross product between two vectors in \code{R^3}.
}
\examples{
\dontrun{
x <- c(0,1,3)
y <- c(2,3,4)
z <- cross.prod(x, y)
}
}
