% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesics.R
\name{dist.to.knn}
\alias{dist.to.knn}
\title{Producess k-NN distance and index matrices associated with a distance matrix.}
\usage{
dist.to.knn(d, k)
}
\arguments{
\item{d}{A distance matrix.}

\item{k}{The number of nearest neighbors.}
}
\value{
A list containing two matrices:
\item{nn.i}{An n x k matrix where element \code{[i,j]} contains the index of the j-th
nearest neighbor of point i.}
\item{nn.d}{An n x k matrix where element \code{[i,j]} contains the distance to the j-th
nearest neighbor of point i.}
}
\description{
Producess k-NN distance and index matrices associated with a distance matrix.
}
\details{
For each point (row) in the distance matrix, this function identifies the k nearest
neighbors and their corresponding distances. The neighbors are ordered from closest
to farthest.
}
\examples{
# Create a simple distance matrix
d <- as.matrix(dist(matrix(rnorm(20), ncol=2)))
knn_result <- dist.to.knn(d, k=3)
# knn_result$nn.i contains indices of 3 nearest neighbors for each point
# knn_result$nn.d contains distances to those neighbors

}
