% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sampling.R
\name{dlaplace}
\alias{dlaplace}
\title{Laplace Distribution Density Function}
\usage{
dlaplace(x, location = 0, scale = 1, log = FALSE)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{location}{The location parameter \eqn{\mu}. Default is 0.}

\item{scale}{The scale parameter b. Must be positive. Default is 1.}

\item{log}{Logical; if TRUE, the log density is returned. Default is FALSE.}
}
\value{
A vector of density values (or log-density if log = TRUE).
}
\description{
Calculates the density function for the Laplace distribution.
}
\examples{
x <- seq(-5, 5, by = 0.1)
y <- dlaplace(x, location = 0, scale = 1)
plot(x, y, type = "l", main = "Laplace Density")

}
