% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{draw.3d.line}
\alias{draw.3d.line}
\title{Draw 3D Line Segment}
\usage{
draw.3d.line(x, length = 2, col = "gray")
}
\arguments{
\item{x}{Numeric vector of length 3 specifying direction.}

\item{length}{Length of the line segment.}

\item{col}{Color of the line segment.}
}
\value{
Invisibly returns NULL.
}
\description{
Plots a line segment in 3D space from origin in the direction of a vector
}
\details{
This function draws a line segment from the origin (0,0,0) in the direction
of vector x, with the specified length.
}
\examples{
\dontrun{
plot3D.plain(matrix(0, ncol = 3))
draw.3d.line(c(1, 1, 1), length = 2, col = "red")
draw.3d.line(c(1, 0, 0), length = 1.5, col = "blue")
}

}
