% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{draw.dashed.line3d}
\alias{draw.dashed.line3d}
\title{Draw Dashed Line in 3D Space}
\usage{
draw.dashed.line3d(
  x1,
  y1,
  z1,
  x2,
  y2,
  z2,
  n = 10,
  dashLength = 0.1,
  gapLength = 0.1,
  col = "black",
  lwd = 2
)
}
\arguments{
\item{x1, y1, z1}{Coordinates of the starting point.}

\item{x2, y2, z2}{Coordinates of the ending point.}

\item{n}{Deprecated parameter kept for compatibility.}

\item{dashLength}{Length of each dash.}

\item{gapLength}{Length of gaps between dashes.}

\item{col}{Color of the line.}

\item{lwd}{Line width.}
}
\value{
Invisibly returns NULL.
}
\description{
Creates a dashed line between two points in 3D space
}
\details{
This function creates a dashed line by drawing multiple short line segments
with gaps between them.
}
\examples{
\dontrun{
# Draw a dashed line from origin to (1, 1, 1)
plot3D.plain(matrix(c(0,0,0,1,1,1), ncol = 3, byrow = TRUE))
draw.dashed.line3d(0, 0, 0, 1, 1, 1, col = "red", lwd = 3)
}

}
