% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{dx.vs.gRf.mod.gSf}
\alias{dx.vs.gRf.mod.gSf}
\title{Given gSf and gRf, dx are found so that corresponding grid(X) has give gSf.
The resulting data is used to build models allowing to construct grid(X) with
the prespecified values of gRf and gSf.}
\usage{
dx.vs.gRf.mod.gSf(
  X,
  min.K = 15,
  med.dK.divf = 10,
  max.dx.C = 5,
  max.dx = NULL,
  min.gRf = 3,
  max.gRf = 20,
  n.gRfs = 20,
  min.gSf = 1,
  max.gSf = 5,
  n.gSfs = 10,
  n.cores = 10,
  verbose = TRUE
)
}
\arguments{
\item{X}{A state space matrix.}

\item{min.K}{The number of nearest neighbors used to estimate median min.K.}

\item{med.dK.divf}{A med.dK division factor for the lower bound of bw and dx to be med.dK/med.dK.divf, where \code{med.dK=median(dK)} with \code{dK=nn.dist[,min.K]}.}

\item{max.dx.C}{A division factor for max.dx estimate such that max.dx = (length of the diagonal in the X encolosing box) / max.dx.C.}

\item{max.dx}{The maximal value of dx.}

\item{min.gRf}{The minimum of gRf.}

\item{max.gRf}{The maximum of gRf.}

\item{n.gRfs}{The number of values of gRf between min.gRf and max.gRf.}

\item{min.gSf}{The minimum of gSf.}

\item{max.gSf}{The maximum of gSf.}

\item{n.gSfs}{The number of values of gSf between min.gSf and max.gSf.}

\item{n.cores}{The number of cores in foreach loop.}

\item{verbose}{A logical flag. Set it to TRUE to report timing of the foreach loop.}
}
\description{
Given gSf and gRf, dx are found so that corresponding grid(X) has give gSf.
The resulting data is used to build models allowing to construct grid(X) with
the prespecified values of gRf and gSf.
}
