% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesics.R
\name{E.geodesic.F}
\alias{E.geodesic.F}
\title{Estimates and the mean prevalence of a factors along a geodesic.}
\usage{
E.geodesic.F(
  subj.factors,
  subjID,
  bch.ids,
  dist.along.bch.geodesic,
  file.prefix
)
}
\arguments{
\item{subj.factors}{A list of subject-level factors.}

\item{subjID}{A vector of subject IDs.}

\item{bch.ids}{Branch IDs.}

\item{dist.along.bch.geodesic}{A distance along the geodesic of the given branch.}

\item{file.prefix}{The prefix of an rda file to which the results of E_gamma_F estimates will be save. Example: "~/projects/rllm/projects/ZB/data/asv_branches/Lc2Gv6_".}
}
\value{
A named character vector mapping factor IDs to their corresponding saved
file paths. Each element name is a factor ID and its value is the full path
to the RDA file containing the E_gamma_F estimation results for that factor.
}
\description{
Given a list of subject-level factors, subj.factors, and IDs of the given
branch, bch.ids, as well as the distance along the geodesic, gamma, vector
and a file prefix, this routine estimates the mean prevalence of the given
factor, E_gamma_F, along the geodesic and saves the results of the estimates
to a file the given prefix and suffix: Fid.rda.
}
\details{
For each factor in subj.factors, this function:
\enumerate{
\item Identifies subjects present in the branch
\item Calculates the mean prevalence along the geodesic distance
\item Saves the results to an RDA file named \verb{\{file.prefix\}\{Fid\}.rda}
}

The saved RDA files contain the E_gamma_F estimates which can be loaded for
further analysis or visualization.
}
