% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{ecdf.cpp}
\alias{ecdf.cpp}
\title{Empirical Cumulative Distribution Function (ECDF)}
\usage{
ecdf.cpp(x)
}
\arguments{
\item{x}{A numeric vector of data.}
}
\value{
A numeric vector representing the ECDF values for each unique value in the input vector.
The length of the return value is the same as the number of unique values in the input vector.
The names attribute of the return value is set to "ecdf".
}
\description{
Calculates the empirical cumulative distribution function (ECDF) for a given vector of data.
}
\details{
The ECDF is a step function that represents the cumulative distribution of a sample of data.
It is defined as the proportion of observations in the sample that are less than or equal to
a given value. This function calculates the ECDF for each unique value in the input vector.
}
\note{
The input vector is sorted internally to compute the ECDF. If the input vector contains
missing values (NA), they will be removed before computing the ECDF.
}
\examples{
x <- c(1.0, 2.0, 3.0, 4.0, 5.0)
result <- ecdf.cpp(x)
print(result)

}
