% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{edge.diff}
\alias{edge.diff}
\title{Compute Edge Difference Between Two Graphs}
\usage{
edge.diff(graph1, graph2)
}
\arguments{
\item{graph1}{A list representing the adjacency list of the first graph.
Each element of the list corresponds to a vertex and contains the indices of its neighbors.}

\item{graph2}{A list representing the adjacency list of the second graph.
It must have the same number of vertices as graph1.}
}
\value{
A list of the same length as the input graphs. Each element contains
the neighbors of the corresponding vertex that are present in graph1 but not in graph2.
}
\description{
This function takes two graphs represented as adjacency lists and computes
the edge difference from graph1 to graph2. It returns a list where each
component contains the edges present in graph1 but not in graph2 for each vertex.
}
\examples{
graph1 <- list(c(2,3), c(1,3), c(1,2), c(5), c(4))
graph2 <- list(c(2), c(1), c(), c(5), c(4))
diff <- edge.diff(graph1, graph2)
print(diff)

}
