% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spectral_ma_lowess.R
\name{effective.models}
\alias{effective.models}
\alias{effective.models.graph.spectral.ma.lowess}
\title{Effective Number of Models in Graph Spectral MA LOWESS}
\usage{
effective.models(object, ...)

\method{effective.models}{graph.spectral.ma.lowess}(object, ...)
}
\arguments{
\item{object}{A 'graph.spectral.ma.lowess' object}

\item{...}{Additional arguments (currently unused)}
}
\value{
Numeric vector with effective number of models per vertex
}
\description{
Calculates the effective number of models contributing to each vertex's prediction
based on the weight distribution entropy.
}
