% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{elapsed.time}
\alias{elapsed.time}
\title{Format and Print Elapsed Time}
\usage{
elapsed.time(start.time, message = "DONE", with.brackets = TRUE)
}
\arguments{
\item{start.time}{A proc.time() object representing the start time.}

\item{message}{Character string to be printed before the elapsed time.
Default is "DONE".}

\item{with.brackets}{Logical, if TRUE (default), the time is enclosed
in parentheses.}
}
\value{
None. This function is called for its side effect of printing.
}
\description{
This function calculates the elapsed time since a given start time,
formats it, and prints it with an optional message.
}
\examples{
\dontrun{
start <- proc.time()
Sys.sleep(2)  # Do some work
elapsed.time(start, "Processing complete")
}
}
