% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{entropy}
\alias{entropy}
\title{Entropy of a vector of non-negative values}
\usage{
entropy(x, base = 10)
}
\arguments{
\item{x}{A vector of non-negative values}

\item{base}{The base of the logarithm in the entropy formula.}
}
\value{
entropy with log 'base'
}
\description{
Entropy of a vector of non-negative values
}
