% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_smooth_morse_smale.R
\name{evaluate.function.on.grid.as.vector}
\alias{evaluate.function.on.grid.as.vector}
\title{Evaluate Function on Grid as Vector}
\usage{
evaluate.function.on.grid.as.vector(f, grid)
}
\arguments{
\item{f}{Function taking two arguments (x, y)}

\item{grid}{List containing grid data (as created by create.grid)}
}
\value{
Numeric vector of function values
}
\description{
Evaluates a bivariate function at each point of a grid, returning results as a vector.
}
\examples{
grid <- create.grid(50)
f <- function(x, y) x^2 + y^2
values <- evaluate.function.on.grid.as.vector(f, grid)
}
