% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_smooth_morse_smale.R
\name{evaluate.function.on.grid}
\alias{evaluate.function.on.grid}
\title{Evaluate Function on Grid}
\usage{
evaluate.function.on.grid(f, grid)
}
\arguments{
\item{f}{Function taking two arguments (x, y)}

\item{grid}{List containing grid data (as created by create.grid)}
}
\value{
Matrix of function values
}
\description{
Evaluates a bivariate function on a uniform grid.
}
\examples{
\dontrun{
grid <- create.grid(50)
f <- function(x, y) sin(x) * cos(y)
f.grid <- evaluate.function.on.grid(f, grid)
image(grid$x, grid$y, f.grid)
}
}
