% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{evenness}
\alias{evenness}
\title{Computes Shannon Evenness of a discrete probability distribution}
\usage{
evenness(x, base = 2)
}
\arguments{
\item{x}{Numeric vector of counts or abundances}

\item{base}{Logarithm base (default: 2)}
}
\value{
A number in \eqn{[0,1]} quantifying evenness (1 = perfectly even)
}
\description{
Computes normalized Shannon entropy (evenness) of a numeric vector defined as entropy(x) / logB(length(x)), where B is the base of the logarithm.
}
