% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{expand.box}
\alias{expand.box}
\title{Expands a bounding box given by the opposite vertices  (L, R)}
\usage{
expand.box(L, R, p.exp, margin = NULL)
}
\arguments{
\item{L}{A numeric vector representing the left vertix that is the point of the box with the smallest coordinates among all other points of the box.}

\item{R}{A numeric vector representing the right vertix that is the point of the box with the largest coordinates among all other points of the box.}

\item{p.exp}{An expansion factor for the bounding box. Each edge of the box is scaled by the factor (1 + p.exp).}

\item{margin}{A numeric value specifying a fixed margin to add to the box. If NULL (default), margin is calculated based on p.exp.}
}
\description{
Expands a bounding box given by the opposite vertices  (L, R)
}
