% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerve_cx.R
\name{extract.skeleton.graph}
\alias{extract.skeleton.graph}
\title{Extract 1-Skeleton Graph from Nerve Complex}
\usage{
extract.skeleton.graph(complex)
}
\arguments{
\item{complex}{A nerve complex object}
}
\value{
A graph representation of the 1-skeleton
}
\description{
This function extracts the 1-skeleton of the nerve complex as a graph.
}
