% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fassoc1_test.R
\name{fassoc1.test}
\alias{fassoc1.test}
\alias{fofam.test}
\title{Tests for First-Order Functional Association Between Two Variables}
\usage{
fassoc1.test(
  x,
  y,
  two.sample.test = c("norm", "Wasserstein", "KS"),
  bw = NULL,
  n.perms = 1000,
  n.BB = 1000,
  grid.size = 400,
  min.K = 5,
  n.cores = 7,
  plot.it = TRUE,
  xlab = "x",
  ylab = "y",
  Eyg.col = "red",
  pt.col = "black",
  pt.pch = 1,
  CrI.as.polygon = TRUE,
  CrI.polygon.col = "gray90",
  null.lines.col = "gray95",
  CrI.line.col = "gray",
  CrI.line.lty = 2,
  verbose = TRUE
)

fofam.test(
  x,
  y,
  two.sample.test = c("norm", "Wasserstein", "KS"),
  bw = NULL,
  n.perms = 1000,
  n.BB = 1000,
  grid.size = 400,
  min.K = 5,
  n.cores = 7,
  plot.it = TRUE,
  xlab = "x",
  ylab = "y",
  Eyg.col = "red",
  pt.col = "black",
  pt.pch = 1,
  CrI.as.polygon = TRUE,
  CrI.polygon.col = "gray90",
  null.lines.col = "gray95",
  CrI.line.col = "gray",
  CrI.line.lty = 2,
  verbose = TRUE
)
}
\arguments{
\item{x}{A numeric vector of predictor values.}

\item{y}{A numeric vector of response values (same length as x). Can be binary or continuous.}

\item{two.sample.test}{Character string specifying the test for comparing distributions.
Options are "Wasserstein", "KS", or "norm". Default is "norm".}

\item{bw}{Numeric bandwidth parameter for smoothing. If NULL (default), bandwidth is
automatically selected.}

\item{n.perms}{Integer specifying the number of permutations for null distribution.
Default is 1000.}

\item{n.BB}{Integer specifying the number of Bayesian bootstrap samples. Default is 1000.}

\item{grid.size}{Integer specifying the size of evaluation grid. Default is 400.}

\item{min.K}{Integer specifying minimum number of observations for local estimation.
Default is 5.}

\item{n.cores}{Integer specifying the number of CPU cores for parallel computation.
Default is 7.}

\item{plot.it}{Logical indicating whether to produce diagnostic plots. Default is TRUE.}

\item{xlab}{Character string for x-axis label. Default is "x".}

\item{ylab}{Character string for y-axis label. Default is "y".}

\item{Eyg.col}{Color specification for conditional mean curve. Default is "red".}

\item{pt.col}{Color specification for data points. Default is "black".}

\item{pt.pch}{Integer or character specifying point type. Default is 1.}

\item{CrI.as.polygon}{Logical indicating whether to draw credible intervals as
polygons (TRUE) or lines (FALSE). Default is TRUE.}

\item{CrI.polygon.col}{Color specification for credible interval polygon.
Default is "gray90".}

\item{null.lines.col}{Color specification for null distribution lines.
Default is "gray95".}

\item{CrI.line.col}{Color specification for credible interval lines.
Default is "gray".}

\item{CrI.line.lty}{Line type for credible interval lines. Default is 2.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}
}
\value{
An object of class "assoc1" containing:
\item{Delta1}{The total change in conditional mean (E_y(x_max) - E_y(x_min))}
\item{delta1}{The first-order functional association measure}
\item{p.value}{The p-value for the test of association}
\item{log.p.value}{Natural logarithm of the p-value}
\item{null.delta1}{Vector of null distribution values}
\item{null.delta1.boxcox}{Box-Cox transformed null distribution (if applicable)}
\item{BB.delta1}{Bayesian bootstrap samples of delta1}
\item{BB.delta1.boxcox}{Box-Cox transformed bootstrap samples (if applicable)}
\item{delta1.boxcox}{Box-Cox transformed observed delta1 (if applicable)}
\item{null.Ey}{Matrix of null conditional mean estimates}
\item{null.dEy}{Matrix of null conditional mean derivatives}
\item{BB.dEy}{Matrix of bootstrap conditional mean derivatives}
\item{Ey.res}{Results from conditional mean estimation}
\item{x}{Input predictor values}
\item{y}{Input response values}
\item{Ey}{Conditional mean estimates at grid points}
\item{grid.size}{Number of grid points used}
}
\description{
Computes and tests the first-order functional association measure between two
variables x and y. The function estimates the integral of |dE_x(y)/dx|, where
E_x(y) is the conditional mean of y given x. This measure captures the strength
of the derivative relationship between variables.
}
\details{
The first-order functional association measure captures the variability in the
derivative of the conditional mean function. Unlike the zero-order measure which
looks at deviations from the mean, this measure is sensitive to the rate of change
in the relationship between x and y. The test is performed by comparing the observed
measure against a null distribution generated through permutation with Bayesian
bootstrap for uncertainty quantification.
}
\examples{
\dontrun{
# Generate example data with nonlinear relationship
set.seed(123)
n <- 200
x <- runif(n)
y <- sin(4*pi*x) + rnorm(n, sd = 0.2)

# Test for first-order functional association
result <- fassoc1.test(x, y, n.cores = 2, n.perms = 500)

# Plot derivative of conditional mean
plot(result, plot = "dExy")
}

}
