% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesics.R
\name{fig.E.geodesic.F}
\alias{fig.E.geodesic.F}
\title{Generates figures associated with E.geodesic.F estimates}
\usage{
fig.E.geodesic.F(save.file.prefix, rlabs, pics.file.prefix)
}
\arguments{
\item{save.file.prefix}{A prefix of the save() files.}

\item{rlabs}{Labels of the subject factors with levels.}

\item{pics.file.prefix}{A prefix pdf output files.}
}
\value{
NULL. This function is called for its side effects of generating PDF plots.
One PDF file is created for each factor in rlabs that has a corresponding saved
RDA file.
}
\description{
Generates figures associated with E.geodesic.F estimates
}
\details{
This function loads the results from E.geodesic.F estimation (saved as RDA files)
and creates visualization plots saved as PDF files. For each factor label in rlabs,
it looks for a file named \verb{\{save.file.prefix\}\{lab\}.rda} and if found, generates
a plot saved as \verb{\{pics.file.prefix\}\{lab\}_Exy.pdf}.

The plots show the expected values (Ey) along the geodesic distance (Ex) with
confidence intervals, providing a visual representation of how the factor varies
along the geodesic path.
}
\seealso{
\code{\link{E.geodesic.F}} for generating the data files used by this function
}
