% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{find.box.containing.x}
\alias{find.box.containing.x}
\title{Finds the box within a box list covering some data set that contains a given point.}
\usage{
find.box.containing.x(boxes, x)
}
\arguments{
\item{boxes}{A list of boxes, where each box is a list containing vectors \code{L} and \code{R} representing the left and right boundaries of the box.}

\item{x}{A vector representing the point for which the containing sub-box is to be found.}
}
\value{
A list representing the sub-box containing \code{x}, or \code{NULL} if \code{x} is not found within any sub-box.
}
\description{
This function finds the sub-box within a given set of boxes that contains a specified point \code{x}.
}
\examples{
\dontrun{
boxes <- create.ED.boxes(0.5, c(0,0), c(1,1))
x <- c(0.2, 0.3)
containing_box <- find.box.containing.x(boxes, x)
}
}
