% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{find.boxes.containing.x}
\alias{find.boxes.containing.x}
\title{Finds boxes within a box list covering some data set that (after expansion by a factor p.exp) contain a given point}
\usage{
find.boxes.containing.x(boxes, x, p.exp = 0.1, margin = NULL)
}
\arguments{
\item{boxes}{A list of boxes, where each box is a list containing vectors \code{L} and \code{R} representing the left and right boundaries of the box.}

\item{x}{A vector representing the point for which the containing sub-box is to be found.}

\item{p.exp}{An expansion factor for the bounding box. Each edge of the box is scaled by the factor (1 + p.exp). Default is 0.1.}

\item{margin}{An explicitly expansion amount of the box in along each axis.}
}
\value{
A list representing the sub-box containing \code{x}, or \code{NULL} if \code{x} is not found within any sub-box.
}
\description{
Finds boxes within a box list covering some data set that (after expansion by a factor p.exp) contain a given point
}
\examples{
\dontrun{
boxes <- create.ED.boxes(0.5, c(0,0), c(1,1))
x <- c(0.2, 0.3)
containing_box <- find.box.containing.x(boxes, x)
}
}
