% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_smooth_morse_smale.R
\name{find.critical.points}
\alias{find.critical.points}
\title{Find Critical Points in Function Grid}
\usage{
find.critical.points(f.grid)
}
\arguments{
\item{f.grid}{Matrix of function values}
}
\value{
List containing:
\item{maxima}{Matrix of (i,j) coordinates for local maxima}
\item{minima}{Matrix of (i,j) coordinates for local minima}
}
\description{
Identifies local maxima and minima in a gridded function.
}
\examples{
\dontrun{
grid <- create.grid(50)
f <- function(x, y) sin(3*x) * cos(3*y)
f.grid <- evaluate.function.on.grid(f, grid)
critical <- find.critical.points(f.grid)
print(critical)
}
}
