% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_smooth_morse_smale.R
\name{find.line.intersection}
\alias{find.line.intersection}
\title{Find Line Intersection}
\usage{
find.line.intersection(P1, P2, Q1, Q2)
}
\arguments{
\item{P1}{First point of first line segment c(x, y)}

\item{P2}{Second point of first line segment c(x, y)}

\item{Q1}{First point of second line segment c(x, y)}

\item{Q2}{Second point of second line segment c(x, y)}
}
\value{
Coordinates of intersection point or NULL if no intersection
}
\description{
Finds the intersection point of two line segments.
}
