% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_MS_cx.R
\name{find.longest.true.stretch}
\alias{find.longest.true.stretch}
\title{Finds the Longest Stretch of Consecutive TRUE Values in a Logical Vector}
\usage{
find.longest.true.stretch(l)
}
\arguments{
\item{l}{A logical vector.}
}
\value{
A list with two elements:
\describe{
\item{start_index}{The index at which the longest stretch of consecutive TRUE values starts.}
\item{length}{The length of the longest stretch of consecutive TRUE values.}
}
}
\description{
This function takes a logical vector as input and returns the starting index
and length of the longest sub-interval where the values are TRUE.
}
\note{
If there are multiple disjoined stretch of TRUE values of the same length, then the first one is returned.
}
\examples{
l <- c(TRUE, TRUE, FALSE, TRUE, TRUE, TRUE, FALSE)
result <- find.longest.true.stretch(l)
print(result)
# $start_index
# [1] 4
# $length
# [1] 3

## Two longest TRUE stretches
l <- c(TRUE, TRUE, TRUE, FALSE, TRUE, TRUE, TRUE, FALSE)
find.longest.true.stretch(l)
# $start.index
# [1] 1
# $length
# [1] 3

## No TRUE values
l <- c(FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE)
find.longest.true.stretch(l)
# $start.index
# [1] NA
# $length
# [1] 0

}
