% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{find.points.within.box}
\alias{find.points.within.box}
\title{Find the points of a set X within a box}
\usage{
find.points.within.box(X, box, eps)
}
\arguments{
\item{X}{A numeric matrix corresponding to a set of points in some Euclidean space.}

\item{box}{A matrix or data frame containing the points in the space.}

\item{eps}{A distance threshold (non-negative real number) such that each box is expanded by eps and then tested for the presence of the elements of X.}
}
\value{
A vector of row names of X corresponding to the points fund within the given box.
}
\description{
This function finds ids of the points of X that are found within the given box.
}
\examples{
\dontrun{
X <- matrix(runif(100), ncol = 2)
box <- list()
box$L <- c(0.2, 0.2)
box$R <- c(0.5, 0.5)
points_within_box <- find.points.within.box(X, box, eps = 0)
}

}
