% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_smooth_morse_smale.R
\name{find.separatrices}
\alias{find.separatrices}
\title{Find Separatrices from Saddle Points}
\usage{
find.separatrices(saddle_point, mixture, eps = 0.01, step = 0.01)
}
\arguments{
\item{saddle_point}{Coordinates of saddle point c(x, y)}

\item{mixture}{Object with gradient method}

\item{eps}{Small offset from saddle (default 0.01)}

\item{step}{Step size for trajectory (default 0.01)}
}
\value{
List of separatrix trajectories
}
\description{
Computes separatrices emanating from saddle points.
}
