% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gflow_cx_utils.R
\name{find.vertices.within.hops}
\alias{find.vertices.within.hops}
\title{Find Vertices Within Specified Hop Distance}
\usage{
find.vertices.within.hops(adj.list, start.vertex, max.hops)
}
\arguments{
\item{adj.list}{Adjacency list representation of the graph}

\item{start.vertex}{Integer index of the starting vertex (1-indexed)}

\item{max.hops}{Integer maximum number of hops to search}
}
\value{
Integer vector of vertex indices that are within max.hops of start.vertex
}
\description{
Uses breadth-first search to find all vertices within a specified number of hops
from a starting vertex.
}
\keyword{internal}
