% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_smooth_morse_smale.R
\name{function.contours.plot}
\alias{function.contours.plot}
\title{Plot Function Contours}
\usage{
function.contours.plot(
  grid,
  f.grid,
  main = "Function Contours",
  nlevels = 20,
  xlab = "",
  ylab = "",
  axes = FALSE,
  add = FALSE,
  ...
)
}
\arguments{
\item{grid}{Grid object from create.grid}

\item{f.grid}{Matrix of function values}

\item{main}{Plot title}

\item{nlevels}{Number of contour levels (default 20)}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{axes}{Logical, whether to show axes}

\item{add}{Logical, whether to add to existing plot}

\item{...}{Additional arguments passed to contour}
}
\value{
Invisible NULL
}
\description{
Creates a contour plot of function values on a grid.
}
\examples{
\dontrun{
grid <- create.grid(50)
f <- function(x, y) sin(3*x) * cos(3*y)
f.grid <- evaluate.function.on.grid(f, grid)
function.contours.plot(grid, f.grid)
}
}
