% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{gaussian.mixture.xD}
\alias{gaussian.mixture.xD}
\title{Creates a Synthetic Function with Specified Number of Local Maxima over a subset of R^d, d > 1}
\usage{
gaussian.mixture.xD(
  S,
  X.lmax = NULL,
  n.lmax = 3,
  y.lmax = NULL,
  y.min = 1,
  y.max = 5,
  C = 2,
  q = 1,
  with.partition.of.unity = FALSE
)
}
\arguments{
\item{S}{A matrix or data frame of points over which a the values of the constructed synthetic function will be evaluated.}

\item{X.lmax}{A matrix or data frame of the location of local maxima.}

\item{n.lmax}{The number of local maxima. Used when X.lmax is NULL. If this is the case, a random set of n.lmax rows of S is taken to be X.lmax.}

\item{y.lmax}{A vector of values the function being created supposed to have at the local maxima. A function that will be constructed is designed to have the values at the local maxima specified by y.lmax, but there are no warranty that it is going to be the case.}

\item{y.min}{Minimum y value for local maxima (defaults to 1).}

\item{y.max}{Maximum y value for local maxima (defaults to 5).}

\item{C}{A scaling factor to control the spread of the bump functions. Defaults to 2.}

\item{q}{A q parameter in radial q-exponential Gaussian functions. Controls the steepness of the radial Gaussian function.}

\item{with.partition.of.unity}{Set to TRUE to use a partition of unity.}
}
\value{
A list containing the following components:
\itemize{
\item \code{X.lmax}: A matrix of the locations of the local maxima.
\item \code{y.lmax}: A vector of function values at the local maxima.
\item \code{y}: A vector of values of the created synthetic function.
}
}
\description{
Generates a synthetic function with specified number of local maxima using a
mixture of Gaussians and a partition of unity. This function is useful for
creating complex, non-linear synthetic data for analysis and testing.
}
\details{
The function generates a set of local maxima and their corresponding
y-values. It then constructs a partition of unity and a matrix of radial
Gaussian functions centered at these local maxima. The final synthetic
function is the sum of these radial Gaussians weighted by the partition
of unity.
}
