% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{gaussian.xD.grad}
\alias{gaussian.xD.grad}
\title{Gradient of the Gaussian function in arbitrary dimension}
\usage{
gaussian.xD.grad(x, mu, sigma, C)
}
\arguments{
\item{x}{Numeric vector of input values}

\item{mu}{Numeric vector of means (same length as x)}

\item{sigma}{Numeric vector of standard deviations (same length as x)}

\item{C}{Amplitude/scaling factor}
}
\value{
Numeric vector of gradient values
}
\description{
Gradient of the Gaussian function in arbitrary dimension
}
