% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{gaussian.xD}
\alias{gaussian.xD}
\title{Gaussian function in arbitrary dimension}
\usage{
gaussian.xD(x, mu, sigma, C)
}
\arguments{
\item{x}{Numeric vector of input values}

\item{mu}{Numeric vector of means (same length as x)}

\item{sigma}{Numeric vector of standard deviations (same length as x)}

\item{C}{Amplitude/scaling factor}
}
\value{
Numeric value of the Gaussian function
}
\description{
Gaussian function in arbitrary dimension
}
