% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{generate.1d.circular.gaussian.mixture}
\alias{generate.1d.circular.gaussian.mixture}
\title{Generate a 1D Circular Gaussian Mixture}
\usage{
generate.1d.circular.gaussian.mixture(
  n.points = 100,
  x.knots = c(0, pi, 1.5 * pi),
  y.knots = c(5, 8, 2.5),
  sd.knot = 0.5,
  out.dir = NULL
)
}
\arguments{
\item{n.points}{An integer specifying the number of points to generate along
the circle. Must be positive. Default is 100.}

\item{x.knots}{A numeric vector specifying the positions of the Gaussian
centers on the circle, in radians. Values are automatically wrapped to
\code{[0, 2pi]}. Must have the same length as y.knots.}

\item{y.knots}{A numeric vector specifying the heights of the Gaussian
centers. Must have the same length as x.knots.}

\item{sd.knot}{A positive numeric value specifying the standard deviation
for all Gaussian distributions. Default is 0.5.}

\item{out.dir}{An optional character string specifying the directory to save
the results. If NULL (default), results are not saved to a file.}
}
\value{
A list containing the following elements:
\item{x}{A numeric vector of x values (angles in radians) from 0 to 2*pi.}
\item{y}{A numeric vector of corresponding y values of the mixture function.}
\item{file.name}{A character string with the name of the file where results
are saved (if out.dir is specified).}
\item{params}{A list of all input parameters used to generate the mixture.}
}
\description{
This function generates a smooth, periodic function over a circle by creating
a mixture of Gaussian distributions. The function ensures continuity and
smoothness at all points, including the 0/2pi connection point.
}
\examples{
# Generate a simple circular Gaussian mixture
result <- generate.1d.circular.gaussian.mixture(
  n.points = 200,
  x.knots = c(0, pi, 1.5*pi),
  y.knots = c(5, 8, 2.5),
  sd.knot = 0.3
)

# Plot the result
plot(result$x, result$y, type = "l", xlab = "Angle (radians)", ylab = "Value")

}
