% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{generate.circle}
\alias{generate.circle}
\title{Generate Equally Spaced Points on a Circle without any noise.}
\usage{
generate.circle(n, radius = 1)
}
\arguments{
\item{n}{The desired number of points to generate.}

\item{radius}{The radius of the circle (default is 1).}
}
\value{
A data frame with two columns, \code{x} and \code{y}, containing the coordinates of the generated points.
}
\description{
This function creates a set of points that are equally spaced along a circle. The first and last point are the same.
}
