% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{generate.noisy.circle.embedding}
\alias{generate.noisy.circle.embedding}
\title{Generate Noisy Circle Data in Higher Dimensions}
\usage{
generate.noisy.circle.embedding(
  n,
  dim = 3,
  radius = 1,
  noise = 0,
  type = "random",
  noise.type = "laplace"
)
}
\arguments{
\item{n}{An integer. The number of data points to generate.}

\item{dim}{An integer greater than 2. The dimension of the space to embed the circle in.}

\item{radius}{A positive number. The radius of the circle.}

\item{noise}{A non-negative number. The scale of the noise to add to the data.}

\item{type}{A string, either "uniform" or "random". Determines how angles are generated.}

\item{noise.type}{A string, either "normal" or "laplace". Determines the distribution of the noise.}
}
\value{
A matrix with n rows and dim columns. Each row represents a point in R^dim.
}
\description{
This function generates data points from a noisy circle embedded in a higher-dimensional space.
}
\examples{
# Generate 100 points on a noisy circle in 3D space
data <- generate.noisy.circle.embedding(100, dim = 3, radius = 1, noise = 0.1)

}
