% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{generate.synthetic.function.higher.dim}
\alias{generate.synthetic.function.higher.dim}
\title{Generate a Synthetic Smooth Function in Higher Dimensions}
\usage{
generate.synthetic.function.higher.dim(X, y = NULL)
}
\arguments{
\item{X}{A numeric matrix where each row represents a point in a multi-dimensional space (R^d).}

\item{y}{An optional numeric vector of response values corresponding to each row in \code{X}.
If not provided, random values are generated.}
}
\value{
A function that represents the synthetic smooth function.
This function takes a numeric vector (a point in R^d) as input and returns a numeric value.
}
\description{
This function creates a synthetic smooth function based on a set of points in a multi-dimensional space.
It utilizes Gaussian functions centered at each point with standard deviations determined by the minimum
distance to other points. The function allows for both positive and negative Gaussians.
}
\details{
The function calculates the standard deviation for each Gaussian centered at a point in \code{X}
as half the minimum distance to any other point in \code{X}. It generates a random vector of +1 or -1
to multiply with the Gaussians, allowing the inclusion of negative values. The output is a function
that, when evaluated at any given point in R^d, provides the corresponding value of
the synthetic smooth function.
}
\examples{
\dontrun{
# Example usage
X <- matrix(runif(20), ncol = 2) # Random 2D points
synthetic_fn <- generate_synthetic_function_higher_dim(X)
point <- c(0.5, 0.5) # A point in 2D space
value <- synthetic_fn(point) # Evaluate the function at the given point
}
}
