% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{generate.trefoil.knot}
\alias{generate.trefoil.knot}
\title{Generate a Trefoil Knot in 3D}
\usage{
generate.trefoil.knot(n = 100, scale = 1, type = "uniform")
}
\arguments{
\item{n}{The desired number of points to generate.}

\item{scale}{The scale of the knot.}

\item{type}{A character vector: "uniform" or "random".}
}
\description{
Generate a Trefoil Knot in 3D
}
