% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{generate_2d_centers}
\alias{generate_2d_centers}
\title{Generate 2D component centers}
\usage{
generate_2d_centers(
  strategy,
  n.components,
  x.range,
  y.range,
  edge.buffer,
  custom.centers
)
}
\description{
Generate 2D component centers
}
\keyword{internal}
