% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{generate_amplitudes_2d}
\alias{generate_amplitudes_2d}
\title{Generate amplitudes for 2D mixture}
\usage{
generate_amplitudes_2d(
  strategy,
  n.components,
  amplitude.range,
  custom.amplitudes
)
}
\description{
Generate amplitudes for 2D mixture
}
\keyword{internal}
