% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{generate_covariances_2d}
\alias{generate_covariances_2d}
\title{Generate covariance matrices for 2D mixture}
\usage{
generate_covariances_2d(
  strategy,
  n.components,
  sd.value,
  sd.range,
  custom.sds,
  correlation
)
}
\description{
Generate covariance matrices for 2D mixture
}
\keyword{internal}
