% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{generate_sampling_points}
\alias{generate_sampling_points}
\title{Generate sampling points using different methods}
\usage{
generate_sampling_points(
  n,
  x.range,
  y.range,
  method = c("random", "grid", "stratified")
)
}
\arguments{
\item{n}{Number of points to generate}

\item{x.range}{Numeric vector of length 2 giving x-axis bounds}

\item{y.range}{Numeric vector of length 2 giving y-axis bounds}

\item{method}{Sampling method: "random", "grid", or "stratified"}
}
\value{
Matrix of coordinates (n x 2)
}
\description{
Generate sampling points using different methods
}
\keyword{internal}
