% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{get.3D.rect.grid.NN}
\alias{get.3D.rect.grid.NN}
\title{Creates a rectangular 3D grid and returns a vector of NN's}
\usage{
get.3D.rect.grid.NN(
  n,
  X,
  f = 0.2,
  eSDf = 1.5,
  mst.grid = NULL,
  mode.edge.len = NULL
)
}
\arguments{
\item{n}{The number of uniformly spaced points, seq(min(xi), max(xi), length=n),
on each axis that are the basis of the grid.}

\item{X}{A set of points around which the grid is created if not NULL.}

\item{f}{A fraction of x1 and x2 range that the grid is extended to.}

\item{eSDf}{A scaling factor, such that edges of mstree(X) are
subdivided is their length is greater than eSDf * mode(edge.len).}

\item{mst.grid}{A matrix of 3D points that are subdivisions of edges of the mstree(X).}

\item{mode.edge.len}{The mode of the edge lengths.}
}
\description{
Creates a rectangular 3D grid and returns a vector of NN's
}
