% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerve_cx.R
\name{get.simplex.counts}
\alias{get.simplex.counts}
\title{Get Simplex Counts for Nerve Complex}
\usage{
get.simplex.counts(complex)
}
\arguments{
\item{complex}{A nerve complex object}
}
\value{
Vector of simplex counts
}
\description{
This function returns the number of simplices in each dimension.
}
