% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesics.R
\name{get.subj.ids}
\alias{get.subj.ids}
\title{Get subject-specific visit information and indices}
\usage{
get.subj.ids(sID, S.3d, subjID, visit)
}
\arguments{
\item{sID}{A subject ID to search for}

\item{S.3d}{A matrix/array with rownames containing subject identifiers}

\item{subjID}{A vector of subject IDs corresponding to rows in S.3d}

\item{visit}{A vector of visit numbers corresponding to subjID entries}
}
\value{
A list containing:
\item{sVisit}{Named vector of visit numbers for the subject}
\item{ii}{Indices of the subject's rows in S.3d}
\item{ids}{Character vector of row names from S.3d for this subject}
}
\description{
Extracts visit information and corresponding row indices in S.3d for a specific subject.
}
\examples{
# Example usage:
# S.3d <- matrix(1:12, nrow=4, dimnames=list(c("s1_v1", "s1_v2", "s2_v1", "s2_v2")))
# subjID <- c("s1", "s1", "s2", "s2")
# visit <- c(1, 2, 1, 2)
# get.subj.ids("s1", S.3d, subjID, visit)
}
