% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{ggaussian}
\alias{ggaussian}
\title{Generalized Gaussian Function}
\usage{
ggaussian(x, offset = 0, h = 1, p = 2)
}
\arguments{
\item{x}{A numeric vector or a single numeric value where the generalized Gaussian function is evaluated.}

\item{offset}{The center of the Gaussian function (defaults to 0).}

\item{h}{The scale parameter of the Gaussian function (defaults to 1).}

\item{p}{The power to which the absolute difference is raised (defaults to 2).}
}
\value{
A numeric vector or a single numeric value representing the value(s) of the generalized Gaussian
function at the input \code{x}. The function's shape changes depending on the value of \code{p}.
}
\description{
Computes the values of a generalized Gaussian function, which allows for adjusting the shape of the curve
by altering the power of the absolute difference term.
}
\details{
The generalized Gaussian function is defined as exp(-((abs(x - offset))^p) / h^p).
The parameter \code{p} allows for adjusting the shape of the Gaussian curve:
a value of \code{p} = 2 gives the standard Gaussian, \code{p} < 2 gives a curve with heavier tails,
and \code{p} > 2 gives a curve with lighter tails. This function can be useful in various
statistical applications where different shapes of Gaussian-like curves are required.
}
\examples{
\dontrun{
x <- seq(-5, 5, length.out = 100)
plot(x, ggaussian(x, p = 2), type = "l")  # Standard Gaussian
plot(x, ggaussian(x, p = 1), type = "l")  # Laplace distribution (heavier tails)
plot(x, ggaussian(x, p = 3), type = "l")  # Lighter tails than Gaussian
}
}
