% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{ggraph}
\alias{ggraph}
\title{Create a gflow graph object}
\usage{
ggraph(adj.list, weight.list = NULL)
}
\arguments{
\item{adj.list}{List where element i contains indices of vertices adjacent to vertex i}

\item{weight.list}{List where element i contains weights of edges from vertex i}
}
\value{
An object of class "ggraph"
}
\description{
Create a gflow graph object
}
