% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{graph.3d}
\alias{graph.3d}
\title{Create a graph.3d object}
\usage{
graph.3d(plot.result, z)
}
\arguments{
\item{plot.result}{A list returned by plot.graph()/plot.ggraph(), containing
the graph and layout elements}

\item{z}{A numeric vector containing z values for each vertex in the graph}
}
\value{
An object of class "graph.3d"
}
\description{
Create a graph.3d object
}
