% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{graph.connected.components}
\alias{graph.connected.components}
\title{Assign Vertices to Connected Components}
\usage{
graph.connected.components(adj.list)
}
\arguments{
\item{adj.list}{A list representing the adjacency list of the graph.
Each element of the list corresponds to a vertex and contains the indices of its neighbors.}
}
\value{
An integer vector where each element represents the connected component
ID for the corresponding vertex.
}
\description{
This function takes a graph representation and assigns each vertex to a connected component.
}
\examples{
adj.list <- list(c(2,3), c(1), c(1), c(5), c(4))
components <- graph.connected.components(adj.list)
print(components)

}
