% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein_dist.R
\name{greedy.NN.transport.distance.1D}
\alias{greedy.NN.transport.distance.1D}
\title{A function to compute the Wasserstein distance between two samples from one-dimensional distributions using nearest neighbors strategy.}
\usage{
greedy.NN.transport.distance.1D(x, y)
}
\arguments{
\item{x}{A sample from a one-dimensional distribution.}

\item{y}{A sample from a one-dimensional distribution.}
}
\description{
The samples are assumed to be of the same size.
}
