% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{hgrid}
\alias{hgrid}
\title{Construct a Hierarchical Uniform Grid Around a State Space}
\usage{
hgrid(
  X,
  w,
  epsilon,
  p.exp = 0.05,
  K = 10,
  n.segments.per.axis = 10,
  n.itrs = 1,
  verbose = TRUE
)
}
\arguments{
\item{X}{A numeric matrix representing the state space for which the grid is to be created.}

\item{w}{The desired edge length of the grid. Points in the grid are spaced at intervals of w.}

\item{epsilon}{The maximum allowed distance from the closest point of X for points in the grid.}

\item{p.exp}{A numeric value representing the expansion factor of the bounding box. Must be between 0 and 1. Default is 0.05.}

\item{K}{The number of nearest neighbors to estimate the denstiy of X at the given point.}

\item{n.segments.per.axis}{An integer number specifying the number of subintervals (of equal
length within each axis, but of potentially of different lengths between
different axes) of each axis that will be used to construct box tiling of
the given box.}

\item{n.itrs}{The number of iterations of box sub-divisions. Default 1.}

\item{verbose}{Set to TRUE for progress messages.}
}
\value{
A numeric matrix containing the coordinates of the points in the hierarchical uniform grid.
Each row represents a point in the grid, and the columns correspond to the dimensions of the space.
}
\description{
This function constructs a uniform grid G(X, w, epsilon) of width
w in the k-dimensional Euclidean space R^k. The grid consists of
points that are not more than epsilon away from the closest point of
the given state space X.
}
\details{
The algorithm first constructs a box tiling of X. For each box, elements of X
that are within that box are identified. The order in which the boxes are
tested for the presence and identity of elements of X is done starting from
the box with the highest density of X elements that is estimated using the
distance, dK, to the K-th nearest neighbor. If a box has elements of X in it,
it is added to a list of boxes-with-elements-of-X and the corresponding
elements of X are deleted from dK. The process is iterated until dK is empty.

This process can be iterated, by performing a second round of rough
subdivisions of the boxes in the list of boxes-with-elements-of-X.

When the process is finished a uniform grid of width w is created within each
box from the list of boxes-with-elements-of-X.
}
\examples{
\dontrun{
# Generate a synthetic 2D state space
X <- matrix(runif(200), ncol = 2)
# Construct the hierarchical grid with edge length 0.1 and epsilon 0.2
grid <- hgrid(X, w = 0.1, epsilon = 0.2)
plot(X, col = "red")
points(grid, col = "blue", pch = 3)
}
}
