% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_utils.R
\name{hist1}
\alias{hist1}
\title{Histogram with customized default settings}
\usage{
hist1(x, main = "", n.breaks = 100, col = "red", ...)
}
\arguments{
\item{x}{A numeric vector of values for which the histogram is plotted.}

\item{main}{The title of the plot. Default is an empty string.}

\item{n.breaks}{The number of breaks (bins) in the histogram. Default is 100.}

\item{col}{The color of the histogram bars. Default is "red".}

\item{...}{Additional parameters passed to \code{\link[graphics]{hist}}.}
}
\value{
Invisibly returns a list of class "histogram" as returned by \code{\link[graphics]{hist}}.
}
\description{
Creates a histogram with customized default settings for n.breaks, color, and appearance.
}
\examples{
# Basic usage
data <- rnorm(1000)
hist1(data)

# With custom settings
hist1(data, main = "Normal Distribution", n.breaks = 50, col = "blue")

# Using additional parameters
hist1(data, main = "Custom Histogram", n.breaks = 30, col = "green",
      xlab = "Values", ylab = "Frequency")

}
