% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_utils.R
\name{hist2}
\alias{hist2}
\title{Plot two overlapping histograms}
\usage{
hist2(
  x1,
  x2,
  x1.lab,
  x2.lab,
  xlab = "",
  n.x1.breaks = 30,
  n.x2.breaks = 30,
  main = "",
  col.x1 = rgb(173, 216, 230, maxColorValue = 255, alpha = 80),
  col.x2 = rgb(255, 192, 203, maxColorValue = 255, alpha = 80),
  legend.pos = "topright"
)
}
\arguments{
\item{x1}{A numeric vector for the first histogram.}

\item{x2}{A numeric vector for the second histogram.}

\item{x1.lab}{Label for x1 in the legend.}

\item{x2.lab}{Label for x2 in the legend.}

\item{xlab}{Label for the x-axis. Default is an empty string.}

\item{n.x1.breaks}{The number of breaks for the first histogram. Default is 30.}

\item{n.x2.breaks}{The number of breaks for the second histogram. Default is 30.}

\item{main}{The title of the plot. Default is an empty string.}

\item{col.x1}{Color of x1 histogram bars. Default is light blue with transparency.}

\item{col.x2}{Color of x2 histogram bars. Default is light pink with transparency.}

\item{legend.pos}{Position of the legend. Default is "topright".
Must be one of "bottomright", "bottom", "bottomleft", "left",
"topleft", "top", "topright", "right", "center" or "none".}
}
\value{
Invisibly returns a list containing:
\item{xlim}{The x-axis limits used in the plot}
\item{ylim}{The y-axis limits used in the plot}
}
\description{
Creates two overlapping histograms with customizable colors and transparency
for comparing two distributions.
}
\examples{
# Basic usage
x1 <- rnorm(1000, mean = 0, sd = 1)
x2 <- rnorm(1000, mean = 2, sd = 1.5)
hist2(x1, x2, x1.lab = "Group A", x2.lab = "Group B")

# With custom settings
hist2(x1, x2, x1.lab = "Control", x2.lab = "Treatment",
      xlab = "Values", main = "Comparison of Distributions",
      n.x1.breaks = 50, n.x2.breaks = 50)

# Custom colors and no legend
hist2(x1, x2, x1.lab = "A", x2.lab = "B",
      col.x1 = "lightgreen", col.x2 = "lightcoral",
      legend.pos = "none")

}
