% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{hor.error.bar}
\alias{hor.error.bar}
\title{Add Horizontal Error Bar to Plot}
\usage{
hor.error.bar(xmin, xmax, y, dyy = 0.025, lwd = 1, col = "black")
}
\arguments{
\item{xmin}{Left limit of error bar.}

\item{xmax}{Right limit of error bar.}

\item{y}{Y-coordinate for the error bar.}

\item{dyy}{Vertical offset for end caps.}

\item{lwd}{Line width.}

\item{col}{Color of error bar.}
}
\value{
Invisibly returns NULL.
}
\description{
Adds a horizontal error bar to an existing 2D plot
}
\examples{
plot(rnorm(10), 1:10, xlim = c(-3, 3))
hor.error.bar(-1, 1, 5, col = "blue")

}
