% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{hungarian.frobenius.graph.matching}
\alias{hungarian.frobenius.graph.matching}
\title{Hungarian-Frobenius Graph Matching Algorithm}
\usage{
hungarian.frobenius.graph.matching(graph1, graph2)
}
\arguments{
\item{graph1}{A list representing the adjacency list of the first graph.}

\item{graph2}{A list representing the adjacency list of the second graph.}
}
\value{
The similarity score between the two graphs, ranging from 0 to 1, where 0 indicates perfect similarity.
}
\description{
This function implements the Hungarian-Frobenius Graph Matching Algorithm to measure the similarity
between two graphs. It uses the Hungarian algorithm to find the optimal permutation of vertices that
minimizes the cost of matching the vertices between the graphs, and then calculates the Frobenius norm
distance between the distance matrices of the graphs after applying the optimal permutation.
}
