% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{identify_points_within_box}
\alias{identify_points_within_box}
\title{Identify Points Within a Box}
\usage{
identify_points_within_box(X, L, R)
}
\arguments{
\item{X}{A matrix or data frame containing the points in the space.}

\item{L}{A numeric vector specifying the lower bounds of the box.}

\item{R}{A numeric vector specifying the upper bounds of the box.}
}
\value{
A matrix or data frame containing only the points that fall within the specified box.
}
\description{
This function identifies the points within a given box specified by the lower and upper bounds.
}
\examples{
\dontrun{
X <- matrix(runif(100), ncol = 2)
L <- c(0.2, 0.2)
R <- c(0.5, 0.5)
points_within_box <- identify_points_within_box(X, L, R)
}

}
