% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{inv.logit}
\alias{inv.logit}
\title{Inverse logit tranformation x -> 1 / (1 + exp(-x)) from real numbers into the unit interval \((0,1)\).}
\usage{
inv.logit(x)
}
\arguments{
\item{x}{A numerical vector.}
}
\value{
Inverse logit transformed data.
}
\description{
Inverse logit tranformation x -> 1 / (1 + exp(-x)) from real numbers into the unit interval \((0,1)\).
}
