% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_generators.R
\name{join.graphs}
\alias{join.graphs}
\title{Join two graphs at specified vertices}
\usage{
join.graphs(graph1, graph2, i1, i2)
}
\arguments{
\item{graph1}{A list representing the adjacency list of the first graph.}

\item{graph2}{A list representing the adjacency list of the second graph.}

\item{i1}{An integer specifying the index of the vertex in the first graph to be joined.}

\item{i2}{An integer specifying the index of the vertex in the second graph to be joined.}
}
\value{
A list representing the adjacency list of the joined graph.
}
\description{
This function joins two graphs by identifying a vertex from each graph. The resulting graph
combines the vertices and edges from both input graphs, with the specified vertices being
treated as a single vertex in the joined graph.
}
\examples{
\dontrun{
graph1 <- list(c(2, 3), c(1), c(1))
graph2 <- list(c(2), c(1, 3), c(2))
joined_graph <- join.graphs(graph1, graph2, 2, 1)
print(joined_graph)
}
}
