% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernel.fn}
\alias{kernel.fn}
\title{A generic interface for evaluating a kernel over a numeric vector}
\usage{
kernel.fn(x, kernel.str = "normal")
}
\arguments{
\item{x}{A numeric vector.}

\item{kernel.str}{The name (character string) of a kernel that will be evaluated over x.}
}
\description{
A generic interface for evaluating a kernel over a numeric vector
}
