% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{left.asymmetric.bump.fn}
\alias{left.asymmetric.bump.fn}
\title{Creates Left Asymmetric Bump Function}
\usage{
left.asymmetric.bump.fn(x, offset = 0, h = 1, q = 1)
}
\arguments{
\item{x}{A numeric vector or a single numeric value where the bump function is evaluated.}

\item{offset}{The center (offset) of the bump function (defaults to 0).
The function behaves differently to the left and right of this point.}

\item{h}{The radius of the support of the bump function (defaults to 1).
The function is zero outside the interval \eqn{[offset - h, offset + h]}.}

\item{q}{The power to which the Gaussian function to the right of the offset is raised (defaults to 1).}
}
\value{
A numeric vector or a single numeric value representing the value(s) of the asymmetric bump function
at the input \code{x}. The function smoothly approaches zero at the boundaries of its support and has
different behavior to the left of the offset.
}
\description{
This function creates an asymmetric bump function, which is a modification of the standard bump function
with asymmetric behavior to the left of the offset. It is smooth with compact support, primarily used
in mathematical analysis and applications requiring non-symmetric smooth functions.
}
\details{
The asymmetric bump function is defined differently to the left and right of the offset.
For \code{x < offset} and \code{offset - x < h}, it is defined as exp(-1 / h - (x - offset)^2).
For \verb{|x - offset| < h}, it follows the standard bump function exp(-1 / (h - (x - offset)^2)).
It is zero outside the specified interval, maintaining smooth transitions and compact support.
}
\examples{
\dontrun{
x <- seq(-2, 2, length.out = 100)
plot(x, AsymmetricBumpFunction(x), type = "l")
}
}
