% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_edit_distance.R
\name{load.graph.data}
\alias{load.graph.data}
\title{Load Graph Data from RDA Files}
\usage{
load.graph.data(
  k.values,
  prefix,
  suffix = "_NN_graph.rda",
  graph.object.name = "S.graph",
  verbose = TRUE
)
}
\arguments{
\item{k.values}{Numeric vector of k values to process. Must be positive integers.}

\item{prefix}{Character string specifying the path and prefix of the RDA files.}

\item{suffix}{Character string specifying the suffix of the RDA files
(default: "_NN_graph.rda").}

\item{graph.object.name}{Character string specifying the name of the graph
object stored in the RDA files (default: "S.graph").}

\item{verbose}{Logical indicating whether to show progress messages
(default: TRUE).}
}
\value{
A list containing two components:
\item{adj.list}{Named list of adjacency lists, one for each k value}
\item{dist.list}{Named list of distance lists, one for each k value}
}
\description{
Loads pruned graph data from a series of RDA files for different k values
and combines them into a list structure containing adjacency and distance lists.
}
\details{
The function expects RDA files to be named according to the pattern:
\code{paste0(prefix, k, suffix)} for each k value.

Each RDA file should contain an object with the name specified in
\code{graph.object.name}, which must have components \code{pruned_adj_list}
and \code{pruned_dist_list}.
}
\examples{
\dontrun{
# Load graphs for k = 5, 10, 15
k.values <- c(5, 10, 15)
graph.data <- load.graph.data(
  k.values,
  prefix = "path/to/graphs/graph_k",
  suffix = "_NN_graph.rda",
  graph.object.name = "S.graph"
)
}

}
