% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_over_graphs_utils.R
\name{loc.const.vertices}
\alias{loc.const.vertices}
\title{Identifies vertices of a graph at which a numeric function is locally constant}
\usage{
loc.const.vertices(adj.list, y, prec = 1e-08)
}
\arguments{
\item{adj.list}{A list representing the graph adjacency list. Each element of the list
should be an integer vector specifying the neighboring vertex indices
for a given vertex. The vertex indices should be 1-based.}

\item{y}{A numeric vector representing the function values at each vertex of the graph.
The length of \code{y} should be equal to the number of vertices in the graph.}

\item{prec}{A numeric scalar specifying the precision threshold used to determine local
constancy. If the absolute difference between the value at a vertex and any
of its neighbors is greater than this threshold, the vertex is not considered
locally constant.}
}
\value{
An integer vector containing the indices of the vertices at which the function
is locally constant. The vertex indices are 1-based.
}
\description{
This function takes an adjacency list representing a graph, a numeric vector of function
values at each vertex, and a precision threshold. It identifies the vertices at which the
function is considered locally constant, meaning the absolute difference between the function
value at the vertex and its neighbors is within the specified precision threshold.
}
