% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spectral_lowess.R
\name{logLik.graph.spectral.lowess}
\alias{logLik.graph.spectral.lowess}
\title{Extract Log-Likelihood from Graph Spectral LOWESS}
\usage{
\method{logLik}{graph.spectral.lowess}(object, ...)
}
\arguments{
\item{object}{A 'graph.spectral.lowess' object}

\item{...}{Additional arguments (currently unused)}
}
\value{
Log-likelihood value
}
\description{
Computes the log-likelihood assuming normal errors
}
