% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{mae.plot}
\alias{mae.plot}
\title{Plot Median Absolute Error}
\usage{
mae.plot(
  mae.mean,
  mae.mad,
  ylab = "Median Absolute Error",
  xlab = "Number of Nearest Neighbors"
)
}
\arguments{
\item{mae.mean}{Vector of MAE means.}

\item{mae.mad}{Vector of MAE median absolute deviations.}

\item{ylab}{Label for y-axis.}

\item{xlab}{Label for x-axis.}
}
\value{
Invisibly returns NULL.
}
\description{
Creates a plot of Median Absolute Error (MAE) with error bars
}
\details{
This function creates a line plot with error bars showing the median absolute
error across different numbers of nearest neighbors. The error bars represent
+/-0.5 * MAD (median absolute deviation).
}
\examples{
\dontrun{
mae.mean <- c(NA, 0.5, 0.4, 0.35, 0.33, 0.32)
mae.mad <- c(NA, 0.1, 0.08, 0.07, 0.06, 0.06)
mae.plot(mae.mean, mae.mad)
}

}
\seealso{
\code{\link{vert.error.bar}}
}
