% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{map.S.to.X}
\alias{map.S.to.X}
\title{Map Sample IDs to 3D Space}
\usage{
map.S.to.X(S, X, radius = 0.075, col = "red", legend.title = NULL)
}
\arguments{
\item{S}{Vector of sample IDs to highlight.}

\item{X}{Matrix or data.frame with 3 columns representing 3D coordinates.}

\item{radius}{Radius of spheres for highlighted samples.}

\item{col}{Color for highlighted samples.}

\item{legend.title}{Title for the legend.}
}
\value{
Invisibly returns NULL.
}
\description{
Highlights a subset of samples in 3D space with different colors and sizes
}
\details{
This function creates a 3D plot where samples in set S are highlighted
with larger spheres in a specified color, while other samples are shown
in gray.
}
\examples{
\dontrun{
X <- matrix(rnorm(300), ncol = 3)
rownames(X) <- paste0("Sample", 1:100)
S <- paste0("Sample", c(1, 5, 10, 15, 20))

map.S.to.X(S, X, radius = 0.075, col = 'red')
}

}
