% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{matrix.format}
\alias{matrix.format}
\title{Format Numbers in a Matrix with Specified Decimal Places}
\usage{
matrix.format(x, digits = 2, nsmall = 2, ...)
}
\arguments{
\item{x}{Input matrix (numeric)}

\item{digits}{Number of digits to round to (default = 2)}

\item{nsmall}{Minimum number of digits to display after decimal point (default = 2)}

\item{...}{Additional arguments (currently unused)}
}
\value{
A character matrix with all numbers formatted according to specifications
}
\description{
Formats all numeric elements in a matrix to have a specified number of decimal places,
ensuring a consistent display with leading and trailing zeros as needed.
}
\examples{
m <- matrix(c(0.7, 1.235, 0.1, 0.8876), nrow = 2)
matrix.format(m)               # default: 2 digits
matrix.format(m, digits = 3)   # 3 digits
matrix.format(m, nsmall = 3)   # at least 3 decimal places

}
