% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{minmax.normalize}
\alias{minmax.normalize}
\title{Min-Max Normalization}
\usage{
minmax.normalize(x, y.min = 0, y.max = 1)
}
\arguments{
\item{x}{A numeric vector to be normalized.}

\item{y.min}{The minimum value in the normalized range (default is 0).}

\item{y.max}{The maximum value in the normalized range (default is 1).}
}
\value{
A numeric vector where the original values of \code{x} have been
scaled to fall within the range \code{[y.min, y.max]}.
}
\description{
Scales a numeric vector to a specified range \eqn{[\text{y.min}, \text{y.max}]}.
This transformation shifts and rescales the data, so the minimum value
corresponds to y.min and the maximum value to y.max.
}
\examples{
x <- c(1, 2, 3, 4, 5)
minmax.normalize(x, 0, 1)  # Normalizes x to the range \eqn{[0, 1]}

}
