% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{mode.1D}
\alias{mode.1D}
\title{Estimates the mode of a numeric vector}
\usage{
mode.1D(x, min.size = 10, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{min.size}{The minimal number of finite values within x. If length(x) < min.size, then the result is NA.}

\item{...}{Parameters to be passed to density()}
}
\description{
Estimates the mode of a numeric vector
}
