% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{model.errors}
\alias{model.errors}
\title{Create a model.errors object}
\usage{
model.errors(integrals, bb.integrals)
}
\arguments{
\item{integrals}{Numeric vector or matrix of integrated error measures}

\item{bb.integrals}{Numeric matrix containing bootstrap samples of the integrated
error measures, used for uncertainty quantification}
}
\value{
An object of class "model.errors" containing the error integrals and
their bootstrap distributions
}
\description{
Constructor function for creating a model.errors object that stores integrated
error measures and their bootstrap distributions for model evaluation.
}
