% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_smooth_morse_smale.R
\name{morse.smale.cells.plot}
\alias{morse.smale.cells.plot}
\title{Plot Morse-Smale Cells}
\usage{
morse.smale.cells.plot(grid, morse_smale_complex, f.grid = NULL, ...)
}
\arguments{
\item{grid}{Grid object from create.grid}

\item{morse_smale_complex}{Morse-Smale complex object}

\item{f.grid}{Optional matrix of function values for contours}

\item{...}{Additional arguments passed to plot}
}
\value{
Invisible NULL
}
\description{
Visualizes Morse-Smale cell decomposition with critical points.
}
\examples{
\dontrun{
grid <- create.grid(30)
f <- function(x, y) sin(3*x) * cos(3*y)
f.grid <- evaluate.function.on.grid(f, grid)
complex <- create.morse.smale.complex(f.grid)
morse.smale.cells.plot(grid, complex, f.grid)
}
}
