% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{mstree.grid}
\alias{mstree.grid}
\title{Creates a quasi-uniform grid over the edges of the minimal spanning tree of a state space}
\usage{
mstree.grid(X, eSDf = 1.5)
}
\arguments{
\item{X}{A set of points for which minimal spanning tree edge subdivision is
to be created.}

\item{eSDf}{An edge subdivision factor, such that a uniform grid is created over the
given edge of mstree(X) if its length is greater than eSDf *
mode(edge.len).}
}
\description{
For each edge of the the minimal spanning tree of X whose length is greater
than eSDf * m a uniform grid with the distance between consecutive points
equal to m, where m = mode(edge lengths).
}
