% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstree.R
\name{mstree}
\alias{mstree}
\title{C implementation of a Minimal Spanning Tree on a matrix X with > 1 column}
\usage{
mstree(X, i = 1)
}
\arguments{
\item{X}{A numeric matrix or data frame with more than one column.}

\item{i}{An index of the row of X from which the algorithm starts. Default is 1.}
}
\value{
A list containing two elements:
\item{edges}{A matrix of edge indices, where each row represents an edge in the MST.}
\item{edge.lens}{A vector of edge lengths corresponding to the edges in the MST.}
}
\description{
C implementation of a Minimal Spanning Tree on a matrix X with > 1 column
}
