% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{n.zeros}
\alias{n.zeros}
\title{Returns the number of times a numeric vector changes the sign between
consecutive points}
\usage{
n.zeros(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Returns the number of times a numeric vector changes the sign between
consecutive points
}
