% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_graphs.R
\name{new.path.graph}
\alias{new.path.graph}
\title{Constructor for path.graph S3 Class}
\usage{
new.path.graph(adj.list, edge.length.list, hop.list, shortest.paths)
}
\arguments{
\item{adj.list}{List of adjacency lists}

\item{edge.length.list}{List of edge lengths}

\item{hop.list}{List of hop counts}

\item{shortest.paths}{List containing shortest path information}
}
\value{
An object of class "path.graph"
}
\description{
Internal constructor function for creating path.graph objects.
This function should not be called directly by users.
}
\keyword{internal}
