% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{non.NA.values}
\alias{non.NA.values}
\title{Returns a matrix with two columns. The first column is the index of x where
x[i] is not NA and the second column is the largest k such that x[i] == x[i + k].}
\usage{
non.NA.values(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Returns a matrix with two columns. The first column is the index of x where
x[i] is not NA and the second column is the largest k such that x[i] == x[i + k].
}
