% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{normalize.and.inv.logit}
\alias{normalize.and.inv.logit}
\title{Normalize and Apply Inverse Logit Transformation}
\usage{
normalize.and.inv.logit(x, y.min = -3, y.max = 3)
}
\arguments{
\item{x}{A numeric vector to be transformed.}

\item{y.min}{The minimum value of the range for Min-Max normalization (default is -3).}

\item{y.max}{The maximum value of the range for Min-Max normalization (default is 3).}
}
\value{
A numeric vector where each original value in \code{x} has been first normalized
and then transformed via the inverse logit function.
}
\description{
First performs Min-Max normalization on a numeric vector to scale it within a specified range,
and then applies the inverse logit function to each element.
}
\details{
The Min-Max normalization scales the data linearly between \code{y.min} and \code{y.max}. The inverse
logit function (logistic function) then transforms these normalized values into probabilities,
mapping any real-valued number into the range (0, 1).
}
\examples{
\dontrun{
x <- rnorm(10)
normalize.and.inv.logit(x)  # Normalizes x and applies inverse logit
}
}
