% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic_stats.R
\name{overlap.distribution.plot}
\alias{overlap.distribution.plot}
\title{Plot the distribution of overlap values}
\usage{
overlap.distribution.plot(x, radius_idx = NULL)
}
\arguments{
\item{x}{A vertex_geodesic_stats object}

\item{radius_idx}{Index of the radius to display. If NULL, uses the radius with the most composite geodesics.}
}
\value{
Invisibly returns NULL
}
\description{
Plot the distribution of overlap values
}
