% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_cx.R
\name{overlap_distance_matrix}
\alias{overlap_distance_matrix}
\title{Compute Pairwise Overlap Distance Matrix}
\usage{
overlap_distance_matrix(basin_vertices_list)
}
\arguments{
\item{basin_vertices_list}{A list of integer vectors.}
}
\value{
A symmetric matrix where entry \eqn{(i, j) = 1 - (|A \cap B| / \min(|A|, |B|)),}
measuring the overlap distance between vectors i and j.
}
\description{
Compute Pairwise Overlap Distance Matrix
}
