% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{p.cases}
\alias{p.cases}
\title{Proportion of cases in a binary variable y}
\usage{
p.cases(y)
}
\arguments{
\item{y}{A binary variable with 0/1 values.}
}
\description{
Proportion of cases in a binary variable y
}
