% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{partition.of.unity.1D}
\alias{partition.of.unity.1D}
\title{Creates a Partition of Unity in 1D}
\usage{
partition.of.unity.1D(
  x.center,
  x.min = 0,
  x.max = 10,
  n.grid = 400,
  C = 2,
  q = 3,
  compact.supp = FALSE
)
}
\arguments{
\item{x.center}{A numeric vector specifying the center points of the bump functions.}

\item{x.min}{The minimum value of the interval over which the partition is created (defaults to 0).}

\item{x.max}{The maximum value of the interval over which the partition is created (defaults to 10).}

\item{n.grid}{The number of points in the grid over the interval (defaults to 400).}

\item{C}{A scaling factor to control the spread of the bump functions. Defaults to 2.}

\item{q}{A q parameter in q-Gaussian}

\item{compact.supp}{Set to TRUE to use bump functions with compact support.}
}
\value{
A list with the following components
\itemize{
\item \code{U}: A matrix with \code{n.grid} rows and \code{length(x.center)} columns, representing
the values of the bump functions at each grid point. The sum of the column values at
each row should be normalized to 1 to ensure a true partition of unity.
\item \code{U.before}: U before normalization.
\item \code{x.grid}: A uniform grid over the inverval \eqn{[x.min, x.max]}.
\item \code{x.centers}: Sorted in the ascending order x.centers.
\item \code{breaks}: Mid points between x.centers.
}
}
\description{
This function creates a partition of unity in one dimension using bump functions.
It is designed to cover a specified interval with a set of overlapping bump functions,
ensuring that the sum of these functions at any point in the interval equals one.
}
\details{
The function uses asymmetric bump functions at the ends of the interval and
symmetric bump functions for internal points. The spread of each bump function
is determined by the distance to its neighboring centers, scaled by \code{C}.
The function checks for valid numeric input and requires at least two center points.
}
\examples{
\dontrun{
x.center <- seq(1, 9, by = 2)
partition <- partition.of.unity.1D(x.center)
plot(seq(0, 10, length.out = 400), partition[,1], type = "l")
}
}
