% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{partition.of.unity.xD}
\alias{partition.of.unity.xD}
\title{Creates a Partition of Unity in dimensions greater than 1}
\usage{
partition.of.unity.xD(S, X.centers = NULL, n.comp = 3, C = 2, q = 3)
}
\arguments{
\item{S}{A matrix or data frame of points over which the partition of unity will be evaluated.}

\item{X.centers}{A matrix or data frame of the centers of the components of the PoU.}

\item{n.comp}{The number of components of the PoU. Used when X.centers is NULL. In this case a random set of n.comp rows of S is taken to be X.centers.}

\item{C}{A scaling factor to control the spread of the bump functions. Defaults to 2.}

\item{q}{A q parameter in radial q-exponential Gaussian functions. Controls the steepness of the radial Gaussian function.}
}
\value{
A list with three components:
\itemize{
\item \code{U}: A matrix of the values of each component of the PoU over the points of S. The return matrix has nrow(S) rows and n.comp columns.
\item \code{nn.d}: A vector of distances to the nearest neighbor center.
\item \code{nn.i}: A vector of indices of the nearest neighbor.
}
}
\description{
This function creates a partition of unity (PoU) over a subset S of \(R^d)\),
where d is the number of columns of S. The components of the partition of
unity are constructed from q-exponential Gaussian radial functions. If a matrix or data
frame, X.centers, of the centers of the components of the PoU, the centers of
the components will be places at these points. If X.centers is NULL, then
the n.comp centers will be selected from the points of S.
}
