% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesics.R
\name{path.dist}
\alias{path.dist}
\title{Distances from the first vertex of a path to each consecutive vertex along
the given path normalized so that the total distance is 1}
\usage{
path.dist(s, V, edge.col = "gray")
}
\arguments{
\item{s}{A sequences of vertex indices of the graph - the path.}

\item{V}{A vertex positions matrix.}

\item{edge.col}{A color of the edges.}
}
\value{
A numeric vector of length equal to the path length, containing the
cumulative distances from the first vertex to each vertex along the path,
normalized to the range \eqn{[0, 1]}. The first element is always 0 and
the last element is always 1.
}
\description{
Distances from the first vertex of a path to each consecutive vertex along
the given path normalized so that the total distance is 1
}
\details{
This function calculates the cumulative Euclidean distances along a path defined
by a sequence of vertex indices. The distances are then normalized by the total
path length so that the result represents relative positions along the path.
}
\note{
The parameter edge.col is currently not used in the function implementation.
}
\examples{
# Create a simple vertex matrix
V <- matrix(c(0,0, 1,0, 1,1, 0,1), ncol=2, byrow=TRUE)
path <- c(1, 2, 3, 4)
distances <- path.dist(path, V)
# distances will be c(0.0, 0.333..., 0.666..., 1.0)

}
