% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{path.fn}
\alias{path.fn}
\title{Animate Path Along Trajectory}
\usage{
path.fn(
  i,
  Epath,
  radius = 8e-04,
  sphere.col = "red",
  path.col = "blue",
  path.lwd = 5
)
}
\arguments{
\item{i}{Index in the Epath matrix.}

\item{Epath}{Matrix of smoothed trajectory positions.}

\item{radius}{Radius of the animated sphere.}

\item{sphere.col}{Color of the animated sphere.}

\item{path.col}{Color of the path trail.}

\item{path.lwd}{Width of the path trail.}
}
\value{
Invisibly returns NULL.
}
\description{
Function for animation of a smoothed trajectory
}
\details{
This function is designed to be used in animation loops. It draws a sphere
at the current position and adds a line segment from the previous position.
Note: This function assumes 'sphere.id' exists in the calling environment.
}
\examples{
\dontrun{
# This function is typically used within an animation loop
Epath <- matrix(rnorm(30), ncol = 3)
sphere.id <- NULL
for (i in 2:nrow(Epath)) {
  path.fn(i, Epath)
  Sys.sleep(0.1)
}
}

}
