% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{persistent.values}
\alias{persistent.values}
\title{Returns a matrix with two columns. The first column is the index of x where
\code{x[i] = x[i+1]} and the second column is the largest k such that \code{x[i] = x[i + k]}}
\usage{
persistent.values(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Returns a matrix with two columns. The first column is the index of x where
\code{x[i] = x[i+1]} and the second column is the largest k such that \code{x[i] = x[i + k]}
}
