% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sampling.R
\name{plaplace}
\alias{plaplace}
\title{Laplace Distribution Function}
\usage{
plaplace(q, location = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{Vector of quantiles.}

\item{location}{The location parameter \eqn{\mu}. Default is 0.}

\item{scale}{The scale parameter b. Must be positive. Default is 1.}

\item{lower.tail}{Logical; if TRUE (default), probabilities are \eqn{P[X \leq x]},
otherwise, \eqn{P[X > x]}.}

\item{log.p}{Logical; if TRUE, probabilities p are given as \eqn{\log(p)}. Default is FALSE.}
}
\value{
A vector of cumulative probabilities.
}
\description{
Calculates the cumulative distribution function for the Laplace distribution.
}
\examples{
x <- seq(-5, 5, by = 0.1)
y <- plaplace(x, location = 0, scale = 1)
plot(x, y, type = "l", main = "Laplace CDF")

}
