% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fassoc1_test.R
\name{plot.assoc1}
\alias{plot.assoc1}
\title{Plot Method for assoc1 Objects}
\usage{
\method{plot}{assoc1}(
  x,
  plot = c("Exy", "dExy", "d1hist", "bc.d1hist"),
  xlab = "x",
  ylab = "y",
  x1.lab = "null.fassoc1",
  x2.lab = "BB.fassoc1",
  show.pval = TRUE,
  show.vlines = FALSE,
  Eyg.col = "blue",
  Eyg.lwd = 2,
  pt.col = "black",
  pt.pch = 1,
  CrI.as.polygon = TRUE,
  CrI.polygon.col = "gray90",
  null.line.col = "gray70",
  null.CrI.col = "gray90",
  CrI.line.col = "gray",
  CrI.line.lty = 2,
  dEy.CrI.col = "cornflowerblue",
  d1hist.x = NULL,
  d1hist.y = NULL,
  title = "",
  ylim = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class "assoc1" from fassoc1.test().}

\item{plot}{Character string specifying plot type. Options are:
\describe{
\item{"Exy"}{Conditional mean function with credible intervals}
\item{"dExy"}{Derivative of conditional mean with credible intervals}
\item{"d1hist"}{Histogram of null distribution with observed value}
\item{"bc.d1hist"}{Histogram of Box-Cox transformed null distribution}
}}

\item{xlab}{Character string for x-axis label. Default is "x".}

\item{ylab}{Character string for y-axis label. Default is "y".}

\item{x1.lab}{Character string for first histogram label. Default is "null.fassoc1".}

\item{x2.lab}{Character string for second histogram label. Default is "BB.fassoc1".}

\item{show.pval}{Logical indicating whether to show p-value. Default is TRUE.}

\item{show.vlines}{Logical indicating whether to show vertical reference lines. Default is FALSE.}

\item{Eyg.col}{Color specification for conditional mean curve. Default is "blue".}

\item{Eyg.lwd}{Line width for conditional mean curve. Default is 2.}

\item{pt.col}{Color specification for data points. Default is "black".}

\item{pt.pch}{Integer or character specifying point type. Default is 1.}

\item{CrI.as.polygon}{Logical indicating whether to draw credible intervals as
polygons (TRUE) or lines (FALSE). Default is TRUE.}

\item{CrI.polygon.col}{Color specification for credible interval polygon.
Default is "gray90".}

\item{null.line.col}{Color specification for null hypothesis line.
Default is "gray70".}

\item{null.CrI.col}{Color specification for null distribution credible intervals.
Default is "gray90".}

\item{CrI.line.col}{Color specification for credible interval lines.
Default is "gray".}

\item{CrI.line.lty}{Line type for credible interval lines. Default is 2.}

\item{dEy.CrI.col}{Color specification for derivative credible intervals.
Default is "cornflowerblue".}

\item{d1hist.x}{Numeric x-coordinate for p-value label in histogram plots.
If NULL, automatically positioned.}

\item{d1hist.y}{Numeric y-coordinate for p-value label in histogram plots.
If NULL, automatically positioned.}

\item{title}{Character string for plot title. Default is "".}

\item{ylim}{Numeric vector of length 2 giving y-axis limits. If NULL,
automatically determined.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisible NULL. Called for side effect of creating plot.
}
\description{
Creates diagnostic plots for objects of class "assoc1" produced by fassoc1.test().
}
\examples{
\dontrun{
# Generate example data
set.seed(123)
n <- 200
x <- runif(n)
y <- sin(4*pi*x) + rnorm(n, sd = 0.2)
result <- fassoc1.test(x, y, n.cores = 2, n.perms = 500, plot.it = FALSE)

# Create different plots
plot(result, plot = "Exy")
plot(result, plot = "dExy")
plot(result, plot = "d1hist")
}

}
