% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_cx.R
\name{plot.basin_cx}
\alias{plot.basin_cx}
\title{Plot method for basin_cx objects}
\usage{
\method{plot}{basin_cx}(x, y, ..., type = c("graph", "basins", "cells", "comparison"))
}
\arguments{
\item{x}{A basin_cx object}

\item{y}{Not used}

\item{...}{Additional arguments passed to specific plotting functions}

\item{type}{Type of plot: "graph" (default), "basins", "cells", or "comparison"}
}
\description{
Plot method for basin_cx objects
}
