% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{plot.box.tiling}
\alias{plot.box.tiling}
\title{Plots boxes of a box tiling of a 2D state space}
\usage{
\method{plot}{box.tiling}(
  x,
  L = NULL,
  R = NULL,
  with.box.ids = TRUE,
  BBox.col = "red",
  col = "gray",
  xlab = "",
  ylab = "",
  sleep = 0,
  ...
)
}
\arguments{
\item{x}{A list of sub-boxes, as returned by \code{create.ED.boxes}.}

\item{L}{The lower left vertex of the bounding box  whose grid is plotted. Default is NULL.}

\item{R}{The upper right vertex of the bounding box  whose grid is plotted. Default is NULL.}

\item{with.box.ids}{Set to TRUE if box IDs are to be displayed.}

\item{BBox.col}{A color to be used for the edges of bounding box whose grid is plotted. Default is "red".}

\item{col}{A color to be used for the edges of sub-boxes. Default is "gray".}

\item{xlab}{Label for x-axis. Default is "".}

\item{ylab}{Label for y-axis. Default is "".}

\item{sleep}{The number of seconds to sleep before drawing the next box.}

\item{...}{Additional graphical parameters to be passed to the plotting functions.}
}
\description{
This function creates a plot showing the edges of boxes of the given box
tiling and their indices in the center of each box. It is designed to work
with the output of the \code{box.tiling} function.
}
\examples{
\dontrun{
w <- 1
L <- c(0, 0)
R <- c(2, 3)
grid_boxes <- create.ED.boxes(w, L, R)
plot.boxes(grid_boxes, L, R)
}

}
