% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{plot.gaussian_mixture}
\alias{plot.gaussian_mixture}
\title{Plot method for gaussian_mixture objects}
\usage{
\method{plot}{gaussian_mixture}(x, n.grid = 50, main = "2D Gaussian Mixture", ...)
}
\arguments{
\item{x}{A gaussian_mixture object}

\item{n.grid}{Number of grid points in each dimension. Default is 50.}

\item{main}{Main title for the plots. Default is "2D Gaussian Mixture"}

\item{...}{Additional arguments (currently unused)}
}
\value{
Invisible NULL
}
\description{
Creates a 2x2 layout with contour plot, 3D surface, heat map, and filled contours.
}
