% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spectral_filter.R
\name{plot.graph_spectral_filter}
\alias{plot.graph_spectral_filter}
\title{Plot Method for Graph Spectral Filter Results}
\usage{
\method{plot}{graph_spectral_filter}(x, which = c(1L, 4L), main = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{"graph_spectral_filter"}}

\item{which}{Integer or character vector specifying which plots to produce:
1 = "gcv" (GCV scores), 2 = "spectrum" (eigenvalue spectrum),
3 = "filter" (filter response), 4 = "signal" (original vs filtered signal)}

\item{main}{Main title for the plots}

\item{...}{Further graphical parameters}
}
\value{
Invisibly returns the input object
}
\description{
Plot Method for Graph Spectral Filter Results
}
