% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{plot.gridX}
\alias{plot.gridX}
\title{Plots the grid of a set around which the grid was created}
\usage{
\method{plot}{gridX}(x, with.bounding.box.pts = TRUE, ...)
}
\arguments{
\item{x}{An output from v1.create.X.grid.xD() or create.X.grid.xD().}

\item{with.bounding.box.pts}{Set to TRUE to show the bounding box.}

\item{...}{Additional arguments passed to the plotting functions.}
}
\description{
Plots the grid of a set around which the grid was created
}
