% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgmalo.R
\name{plot.pgmalo}
\alias{plot.pgmalo}
\title{Plot Method for pgmalo Objects}
\usage{
\method{plot}{pgmalo}(x, type = "diagnostic", ...)
}
\arguments{
\item{x}{An object of class "pgmalo".}

\item{type}{Character string specifying plot type (default: "diagnostic").}

\item{...}{Additional arguments passed to plotting functions.}
}
\value{
NULL (invisibly). Called for side effect of creating plots.
}
\description{
Creates diagnostic and visualization plots for pgmalo model fits.
}
\details{
Currently only implements diagnostic plot showing cross-validation errors
across different h values. For more detailed plotting options, use
\code{\link{plot.upgmalo}} with upgmalo objects.
}
\examples{
# See examples in pgmalo()

}
