% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic_stats.R
\name{plot.vertex_geodesic_stats}
\alias{plot.vertex_geodesic_stats}
\title{Plot method for vertex_geodesic_stats objects}
\usage{
\method{plot}{vertex_geodesic_stats}(x, ...)
}
\arguments{
\item{x}{A vertex_geodesic_stats object}

\item{...}{Additional arguments passed to plot}
}
\value{
Invisibly returns NULL
}
\description{
Plot method for vertex_geodesic_stats objects
}
