% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot2D.tree}
\alias{plot2D.tree}
\title{Add Minimal Spanning Tree to 2D Plot}
\usage{
plot2D.tree(X, T.edges, col = "gray")
}
\arguments{
\item{X}{Matrix with at least 2 columns. Only first 2 columns are used.}

\item{T.edges}{Matrix with 2 columns containing start and end indices of edges.}

\item{col}{Color of the edges.}
}
\value{
Invisibly returns NULL.
}
\description{
Adds edges of a minimal spanning tree to an existing 2D plot
}
\examples{
\dontrun{
X <- matrix(rnorm(20), ncol = 2)
T.edges <- matrix(c(1,2, 2,3, 3,4, 4,5), ncol = 2, byrow = TRUE)

plot(X, pch = 19)
plot2D.tree(X, T.edges)
}

}
