% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{plot3d.box.tiling}
\alias{plot3d.box.tiling}
\title{Plots boxes of a box tiling of a 3D state space.}
\usage{
\method{plot3d}{box.tiling}(x, open3d = TRUE, col = "gray", ...)
}
\arguments{
\item{x}{A list of sub-boxes, as returned by \code{create.ED.boxes}.}

\item{open3d}{Logical. Whether to open a new 3D plotting device (default = TRUE).}

\item{col}{Character. Color of the box edges (default = "gray").}

\item{...}{Additional arguments passed to rgl plotting functions.}
}
\description{
This function creates a 3D visualization of box tilings using the rgl package.
}
