% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot3D.cl}
\alias{plot3D.cl}
\title{Plot a Specific Cluster with Custom Colors}
\usage{
plot3D.cl(cl, cltr, X, ...)
}
\arguments{
\item{cl}{Cluster ID(s) to highlight. Can be a single value or vector.}

\item{cltr}{Vector of cluster IDs.}

\item{X}{A matrix or data.frame with 3 columns representing 3D coordinates.}

\item{...}{Additional arguments passed to \code{\link{plot3D.cltrs}}.}
}
\value{
Invisibly returns the output from plot3D.cltrs.
}
\description{
Creates a 3D plot highlighting one or more specific clusters
}
\details{
This function creates a 3D plot where specified clusters are highlighted with
distinct colors while other clusters are shown in gray shades. If multiple
clusters are specified, they are colored using the mclust color scheme.
}
\examples{
\dontrun{
X <- matrix(rnorm(300), ncol = 3)
cltr <- sample(1:5, 100, replace = TRUE)

# Highlight single cluster
plot3D.cl(2, cltr, X)

# Highlight multiple clusters
plot3D.cl(c(2, 4), cltr, X)
}

}
