% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot3D.diskEmbdg}
\alias{plot3D.diskEmbdg}
\title{Plot 3d Disk Embedding Object}
\usage{
plot3D.diskEmbdg(
  ebdg.obj,
  col = "blue",
  radius = 0.005,
  vertex.radius = 0.02,
  vertex.col = "red",
  vertex.cex = 1.2,
  edge.col = "gray",
  adj.df = NULL
)
}
\arguments{
\item{ebdg.obj}{A diskEmbdg object containing embedding information.}

\item{col}{Color of spheres representing data points.}

\item{radius}{Radius of data point spheres.}

\item{vertex.radius}{Radius of vertex spheres.}

\item{vertex.col}{Color of vertex spheres.}

\item{vertex.cex}{Character expansion factor for vertex labels.}

\item{edge.col}{Color of edges.}

\item{adj.df}{Matrix of text adjustment values for vertex labels.}
}
\value{
Invisibly returns NULL.
}
\description{
Visualizes a disk embedding object in 3D space
}
\details{
This function visualizes a disk embedding where data is projected onto a
sphere with axes represented as points on the sphere and one at the center.
}
\examples{
\dontrun{
# Assuming ebdg.obj is a properly formatted disk embedding object
plot3D.diskEmbdg(ebdg.obj, col = "blue", vertex.col = "red")
}

}
