% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot3D.geodesic}
\alias{plot3D.geodesic}
\title{Plot Geodesic Path in 3D}
\usage{
plot3D.geodesic(S.3d, i1, i2, G, X, edge.col = "gray")
}
\arguments{
\item{S.3d}{Matrix of 3D vertex positions.}

\item{i1}{Index of the starting vertex.}

\item{i2}{Index of the ending vertex.}

\item{G}{An igraph object representing the graph structure.}

\item{X}{Original state space (currently unused but kept for compatibility).}

\item{edge.col}{Color of the geodesic path.}
}
\value{
Invisibly returns the sequence of vertex indices in the shortest path.
}
\description{
Plots the shortest path between two vertices in a graph
}
\details{
This function finds the shortest path between two vertices using igraph's
shortest_paths function and visualizes it in 3D space.
}
\examples{
\dontrun{
library(igraph)
# Create a simple graph
G <- make_ring(10)
S.3d <- matrix(rnorm(30), ncol = 3)

plot3D.plain(S.3d)
plot3D.geodesic(S.3d, 1, 5, G, S.3d)
}

}
