% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plotlcor.1D}
\alias{plotlcor.1D}
\title{Plot Output from lcor.1D()}
\usage{
plotlcor.1D(
  r,
  with.CrI = TRUE,
  use.smoothed.lcors = TRUE,
  title = "",
  xlab = "",
  ylab = "Local Pearson correlation",
  lcor.col = "blue",
  CrI.as.polygon = TRUE,
  CrI.polygon.col = "gray95",
  CrI.line.col = "gray",
  add.box = FALSE,
  add.x.axis = FALSE,
  add.y.axis = FALSE,
  ...
)
}
\arguments{
\item{r}{An output from lcor.1D() function.}

\item{with.CrI}{A logical parameter. If TRUE, credible intervals are plotted.}

\item{use.smoothed.lcors}{A logical parameter. If TRUE, smoothed local correlations will be used
(if available) instead of raw local correlations.}

\item{title}{The title of the plot.}

\item{xlab}{The x-axis label.}

\item{ylab}{The y-axis label.}

\item{lcor.col}{The color of local correlation line.}

\item{CrI.as.polygon}{A logical value parameter. If TRUE, credible interval region is drawn
as a polygon. Otherwise, credible intervals are drawn as contour lines.}

\item{CrI.polygon.col}{The color of the credible interval polygon region.}

\item{CrI.line.col}{The color of the credible interval upper and lower limit lines.}

\item{add.box}{Set to TRUE if box() is to be called after plot(); useful when axes is FALSE.}

\item{add.x.axis}{Set to TRUE if axis(1) is to be called after plot().}

\item{add.y.axis}{Set to TRUE if axis(2) is to be called after plot().}

\item{...}{Graphics parameters passed to \code{\link[graphics]{plot}}.}
}
\value{
Invisibly returns NULL. The function is called for its side effect of creating a plot.
}
\description{
Creates a visualization of local correlation results from the lcor.1D() function
}
\details{
This function visualizes the output from lcor.1D(), which typically contains local correlation
estimates along a one-dimensional variable. The function can display either raw or smoothed
local correlations, with optional credible intervals shown either as a shaded polygon or
as contour lines.
}
\examples{
\dontrun{
# Assuming 'result' is output from lcor.1D()
plotlcor.1D(result, with.CrI = TRUE, title = "Local Correlations")

# Plot with credible intervals as lines instead of polygon
plotlcor.1D(result, with.CrI = TRUE, CrI.as.polygon = FALSE)
}

}
