% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{prediction.errors}
\alias{prediction.errors}
\title{Create a prediction.errors object}
\usage{
prediction.errors(errors, xvals = NULL)
}
\arguments{
\item{errors}{Numeric vector of prediction errors}

\item{xvals}{Optional numeric vector of x-values corresponding to the errors.
Default is NULL.}
}
\value{
An object of class "prediction.errors" containing the errors with
xvals stored as an attribute
}
\description{
Constructor function for creating a prediction.errors object that stores model
prediction errors along with optional x-values for plotting or analysis.
}
