% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fassoc0_test.R
\name{print.assoc0}
\alias{print.assoc0}
\title{Print Method for assoc0 Objects}
\usage{
\method{print}{assoc0}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class "assoc0".}

\item{digits}{Number of significant digits to display.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisible x.
}
\description{
Print Method for assoc0 Objects
}
