% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fassoc1_test.R
\name{print.assoc1}
\alias{print.assoc1}
\title{Print Method for assoc1 Objects}
\usage{
\method{print}{assoc1}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class "assoc1".}

\item{digits}{Number of significant digits to display.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisible x.
}
\description{
Print Method for assoc1 Objects
}
