% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize_circular_graph.R
\name{print.circular_parameterization}
\alias{print.circular_parameterization}
\title{Print Method for circular_parameterization Objects}
\usage{
\method{print}{circular_parameterization}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{"circular_parameterization"} as returned by
\code{\link{parameterize.circular.graph}}.}

\item{digits}{Number of digits to display for numeric values. Default is 4.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Prints a summary of a circular parameterization result.
}
\examples{
# Create example graph
n <- 6
adj.list <- lapply(seq_len(n), function(i) {
  c(if (i == n) 1 else i + 1, if (i == 1) n else i - 1)
})
weight.list <- lapply(adj.list, function(adj) rep(1.0, length(adj)))
result <- parameterize.circular.graph(adj.list, weight.list)
print(result)

}
