% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulogit.R
\name{print.eigen.ulogit}
\alias{print.eigen.ulogit}
\title{Print Method for eigen.ulogit Objects}
\usage{
\method{print}{eigen.ulogit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"eigen.ulogit"}}

\item{...}{Additional arguments (currently ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for eigen.ulogit Objects
}
