% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fassoc_test.R
\name{print.fassoc}
\alias{print.fassoc}
\title{Print Method for Functional Association Tests}
\usage{
\method{print}{fassoc}(x, ...)
}
\arguments{
\item{x}{An object of class "assoc0" or "assoc1".}

\item{...}{Additional arguments passed to specific print methods.}
}
\value{
Invisible x.
}
\description{
Print Method for Functional Association Tests
}
