% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{print.gaussian_mixture}
\alias{print.gaussian_mixture}
\title{Print method for gaussian_mixture objects}
\usage{
\method{print}{gaussian_mixture}(x, ...)
}
\arguments{
\item{x}{A gaussian_mixture object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisible x
}
\description{
Print method for gaussian_mixture objects
}
