% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spectral_ma_lowess.R
\name{print.graph.spectral.ma.lowess}
\alias{print.graph.spectral.ma.lowess}
\title{Print Graph Spectral MA LOWESS Object}
\usage{
\method{print}{graph.spectral.ma.lowess}(x, ...)
}
\arguments{
\item{x}{A 'graph.spectral.ma.lowess' object}

\item{...}{Additional arguments (currently unused)}
}
\value{
Returns x invisibly
}
\description{
Provides a concise summary of a graph spectral model-averaged LOWESS object
}
