% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spectral_filter.R
\name{print.graph_spectral_filter}
\alias{print.graph_spectral_filter}
\title{Print Method for Graph Spectral Filter Results}
\usage{
\method{print}{graph_spectral_filter}(x, digits = 4L, ...)
}
\arguments{
\item{x}{An object of class \code{"graph_spectral_filter"}}

\item{digits}{Integer; number of significant digits to print (default: 4)}

\item{...}{Further arguments passed to or from other methods}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for Graph Spectral Filter Results
}
