% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maelog.R
\name{print.maelog}
\alias{print.maelog}
\title{Print Method for maelog Objects}
\usage{
\method{print}{maelog}(x, ...)
}
\arguments{
\item{x}{A fitted object of class \code{"maelog"}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Prints a brief summary of a fitted \code{maelog} object.
}
\seealso{
\code{\link{maelog}}, \code{\link{summary.maelog}}
}
