% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_maximal_packing.R
\name{print.maximal_packing}
\alias{print.maximal_packing}
\title{Print Method for Maximal Packing Results}
\usage{
\method{print}{maximal_packing}(x, ...)
}
\arguments{
\item{x}{An object of class "maximal_packing"}

\item{...}{Additional arguments (currently ignored)}
}
\value{
Invisible copy of x
}
\description{
Print Method for Maximal Packing Results
}
