% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mknn_graphs.R
\name{print.mknn_graph}
\alias{print.mknn_graph}
\title{Print Method for mknn_graph Objects}
\usage{
\method{print}{mknn_graph}(x, ...)
}
\arguments{
\item{x}{An object of class "mknn_graph".}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns the input object.
}
\description{
Prints a concise summary of a mutual k-nearest neighbor graph object.
}
