% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerve_cx.R
\name{print.nerve_complex}
\alias{print.nerve_complex}
\title{Print Method for Nerve Complex Objects}
\usage{
\method{print}{nerve_complex}(x, ...)
}
\arguments{
\item{x}{A nerve complex object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the object
}
\description{
Print Method for Nerve Complex Objects
}
