% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_maximal_packing.R
\name{print.packing_validation}
\alias{print.packing_validation}
\title{Print Method for Packing Validation Results}
\usage{
\method{print}{packing_validation}(x, ...)
}
\arguments{
\item{x}{An object of class "packing_validation"}

\item{...}{Additional arguments (currently ignored)}
}
\value{
Invisible copy of x
}
\description{
Print Method for Packing Validation Results
}
