% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_graphs.R
\name{print.path.graph.plm}
\alias{print.path.graph.plm}
\title{Print Method for path.graph.plm Objects}
\usage{
\method{print}{path.graph.plm}(x, ...)
}
\arguments{
\item{x}{A path.graph.plm object}

\item{...}{Additional arguments (currently ignored)}
}
\value{
Invisible copy of x
}
\description{
Print a summary of a PLM path graph object.
}
