% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_graphs.R
\name{print.path.graph}
\alias{print.path.graph}
\title{Print Method for path.graph Objects}
\usage{
\method{print}{path.graph}(x, ...)
}
\arguments{
\item{x}{A path.graph object}

\item{...}{Additional arguments (currently ignored)}
}
\value{
Invisible copy of x
}
\description{
Print a concise summary of a path.graph object.
}
\examples{
\dontrun{
pg <- create.path.graph(graph, edge.lengths, h = 2)
print(pg)
}

}
