% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_graphs.R
\name{print.path.graph.series}
\alias{print.path.graph.series}
\title{Print Method for path.graph.series Objects}
\usage{
\method{print}{path.graph.series}(x, ...)
}
\arguments{
\item{x}{A path.graph.series object}

\item{...}{Additional arguments passed to print.path.graph}
}
\value{
Invisible copy of x
}
\description{
Print a summary of a series of path graphs.
}
