% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgmalo.R
\name{print.pgmalo}
\alias{print.pgmalo}
\title{Print Method for pgmalo Objects}
\usage{
\method{print}{pgmalo}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class "pgmalo".}

\item{digits}{Number of significant digits for numeric output (default: 3).}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object invisibly.
}
\description{
Displays a concise summary of a pgmalo model fit.
}
