% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{print.summary.gaussian_mixture}
\alias{print.summary.gaussian_mixture}
\title{Print summary of gaussian_mixture}
\usage{
\method{print}{summary.gaussian_mixture}(x, ...)
}
\arguments{
\item{x}{A summary.gaussian_mixture object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisible x
}
\description{
Print summary of gaussian_mixture
}
