% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gflow_basins.R
\name{print.summary.gflow_basins}
\alias{print.summary.gflow_basins}
\title{Print Summary of Gradient-Flow Basin Results}
\usage{
\method{print}{summary.gflow_basins}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.gflow_basins"}}

\item{...}{Additional arguments (currently ignored)}
}
\description{
Print Summary of Gradient-Flow Basin Results
}
