% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonic_smoother.R
\name{print.summary.harmonic_smoother}
\alias{print.summary.harmonic_smoother}
\title{Print Method for Summary of Harmonic Smoother Results}
\usage{
\method{print}{summary.harmonic_smoother}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.harmonic_smoother"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the input object.
}
\description{
Prints the summary of harmonic smoother results in a formatted manner.
}
\seealso{
\code{\link{summary.harmonic_smoother}}
}
