% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uggmalo.R
\name{print.summary.uggmalo}
\alias{print.summary.uggmalo}
\title{Print Method for UGGMALO Summary}
\usage{
\method{print}{summary.uggmalo}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class "summary.uggmalo"}

\item{digits}{Integer. Number of significant digits to print}

\item{...}{Additional arguments (currently ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for UGGMALO Summary
}
