% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uggmalog.R
\name{print.summary.uggmalog}
\alias{print.summary.uggmalog}
\title{Print method for summary.uggmalog objects}
\usage{
\method{print}{summary.uggmalog}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.uggmalog"}}

\item{...}{Additional arguments (currently ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for summary.uggmalog objects
}
