% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_kmean.R
\name{print.summary.ugkmm}
\alias{print.summary.ugkmm}
\title{Print summary.ugkmm Objects}
\usage{
\method{print}{summary.ugkmm}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class "summary.ugkmm".}

\item{digits}{Number of significant digits.}

\item{...}{Additional arguments (unused).}
}
\description{
Print summary.ugkmm Objects
}
