% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uggmalog.R
\name{print.uggmalog}
\alias{print.uggmalog}
\title{Print method for uggmalog objects}
\usage{
\method{print}{uggmalog}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"uggmalog"}}

\item{...}{Additional arguments (currently ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for uggmalog objects
}
