% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulogit.R
\name{print.ulogit}
\alias{print.ulogit}
\title{Print Method for ulogit Objects}
\usage{
\method{print}{ulogit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ulogit"}}

\item{...}{Additional arguments (currently ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for ulogit Objects
}
