% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgmalo.R
\name{print.upgmalo}
\alias{print.upgmalo}
\title{Print Method for upgmalo Objects}
\usage{
\method{print}{upgmalo}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class "upgmalo".}

\item{digits}{Number of significant digits for numeric output (default: 3).}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object invisibly.
}
\description{
Displays a concise summary of a upgmalo model fit.
}
