% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{project.onto.subspace}
\alias{project.onto.subspace}
\title{Projects a Vector onto a Subspace in R^n}
\usage{
project.onto.subspace(x, U)
}
\arguments{
\item{x}{A numeric vector of length n representing the vector to be projected.}

\item{U}{A matrix of dimensions n x m, where each column is a unit vector
spanning the subspace. The matrix should have full column rank.}
}
\value{
A numeric vector representing the projection of x onto the subspace
spanned by the columns of U.
}
\description{
This function projects a given vector in R^n onto a subspace spanned by
unit vectors. The subspace is defined by a matrix where each column is a
unit vector spanning the subspace.
}
\examples{
u1 <- c(1, 0, 0)
u2 <- c(0, 1, 0)
u3 <- c(0, 0, 1)
U <- matrix(c(u1, u2, u3), nrow=3)
x <- c(2, 3, 4)
project.onto.subspace(x, U)

}
