% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{qexp.gaussian}
\alias{qexp.gaussian}
\title{q-Exponential Gaussian Function}
\usage{
qexp.gaussian(x, offset = 0, h = 1, q = 4)
}
\arguments{
\item{x}{A numeric vector or a single numeric value where the q-Gaussian function is evaluated.}

\item{offset}{The center of the q-Gaussian function (defaults to 0).}

\item{h}{The standard deviation of the Gaussian part of the function (defaults to 1).}

\item{q}{The power to which the Gaussian function is raised (defaults to 4).}
}
\value{
A numeric vector or a single numeric value representing the value(s) of the q-Gaussian function
at the input \code{x}. The function alters the shape of a standard Gaussian curve based on the value of \code{q}.
}
\description{
This function computes the values of a modified Gaussian function, known as a q-Gaussian,
which raises the standard Gaussian to the power of \code{q}. This modification can be used to
adjust the "flatness" or "peakiness" of the Gaussian curve.
}
\details{
The q-Gaussian function is defined as (exp(-((x - offset)^2) / h2))^q.
The parameter \code{q} allows for adjusting the shape of the Gaussian curve:
higher values of \code{q} make the curve peakier, while lower values make it flatter.
This function can be useful in various statistical and signal processing applications
where a modified Gaussian shape is required.
}
\examples{
\dontrun{
x <- seq(-5, 5, length.out = 100)
plot(x, qgaussian(x, q = 4), type = "l")
}
}
