% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sampling.R
\name{qlaplace}
\alias{qlaplace}
\title{Laplace Distribution Quantile Function}
\usage{
qlaplace(p, location = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{Vector of probabilities.}

\item{location}{The location parameter \eqn{\mu}. Default is 0.}

\item{scale}{The scale parameter b. Must be positive. Default is 1.}

\item{lower.tail}{Logical; if TRUE (default), probabilities are \eqn{P[X \leq x]},
otherwise, \eqn{P[X > x]}.}

\item{log.p}{Logical; if TRUE, probabilities p are given as \eqn{\log(p)}. Default is FALSE.}
}
\value{
A vector of quantiles.
}
\description{
Calculates the quantile function for the Laplace distribution.
}
\examples{
p <- seq(0, 1, by = 0.1)
q <- qlaplace(p, location = 0, scale = 1)
plot(p, q, type = "l", main = "Laplace Quantile Function")

}
