% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{quantize.cont.var}
\alias{quantize.cont.var}
\title{Quantize a continuous variable into categories}
\usage{
quantize.cont.var(
  x,
  method = "uniform",
  wins.p = 0.01,
  round = TRUE,
  dig.lab = 2,
  breaks = NULL,
  include.lowest = TRUE,
  start = 1/6,
  end = 0,
  n.levels = 10,
  use.viridis = FALSE,
  brewer.palette = NULL,
  use.unique = FALSE
)
}
\arguments{
\item{x}{Numeric vector.}

\item{method}{"uniform" or "quantile".}

\item{wins.p}{Winsorization proportion (uniform method).}

\item{round}{Round \eqn{[0,1]} endpoints to one decimal (uniform method).}

\item{dig.lab}{Digits for labels.}

\item{breaks}{Custom breaks.}

\item{include.lowest}{Include lowest value in first interval.}

\item{start, end}{Start/end for rainbow().}

\item{n.levels}{Number of levels.}

\item{use.viridis}{Use viridis palette if available.}

\item{brewer.palette}{Name of RColorBrewer palette to use (optional).}

\item{use.unique}{Use only unique x for break calc.}
}
\value{
list(x.cat, x.col.tbl)
}
\description{
Quantize a continuous variable into categories
}
