% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{quantize.for.legend}
\alias{quantize.for.legend}
\title{Quantize Continuous Variable for Legend Display}
\usage{
quantize.for.legend(
  y,
  quantize.method = "uniform",
  quantize.wins.p = 0.02,
  quantize.round = FALSE,
  quantize.dig.lab = 2,
  start = 1/6,
  end = 0,
  n.levels = 10
)
}
\arguments{
\item{y}{A numeric vector to be quantized.}

\item{quantize.method}{Method to use for quantization. Options are "uniform" (equal-width bins)
or "quantile" (equal-frequency bins).}

\item{quantize.wins.p}{Winsorization parameter for the "uniform" method. Values beyond the
p and 1-p quantiles are grouped together. Default is 0.02.}

\item{quantize.round}{Logical. If TRUE, rounds the endpoints of the range.}

\item{quantize.dig.lab}{Number of digits for labels in the cut() function.}

\item{start}{Start value for the color range in HSV color space (0-1).}

\item{end}{End value for the color range in HSV color space (0-1).}

\item{n.levels}{Number of discrete levels for quantization.}
}
\value{
A list containing:
\item{y.cat}{Factor of quantized values}
\item{y.col.tbl}{Named vector mapping categories to colors}
\item{y.cols}{Vector of colors for each input value}
\item{legend.labs}{Formatted labels for legend display}
}
\description{
Quantizes a continuous variable and generates legend colors and labels for visualization
}
\details{
This function is primarily used internally by plotting functions to convert continuous
variables into discrete categories with associated colors for visualization. The
winsorization parameter helps handle outliers in the uniform method.
}
\examples{
\dontrun{
y <- rnorm(100)
q <- quantize.for.legend(y, n.levels = 5)
plot(1:100, y, col = q$y.cols, pch = 19)
legend("topright", legend = q$legend.labs, fill = q$y.col.tbl)
}

}
