% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{radial.Lp.transform}
\alias{radial.Lp.transform}
\title{Perform Lp Radial Transformation}
\usage{
radial.Lp.transform(X, p = 0.5)
}
\arguments{
\item{X}{A matrix or data frame containing numeric values to be transformed.}

\item{p}{The Lp exponent. Must be greater than 0. Default is 0.5.}
}
\value{
A matrix or data frame of the same dimensions as the input, with the Lp radial transformation applied to each row.
}
\description{
This function applies the Lp radial transformation to each row of the input matrix or data frame.
The transformation is defined as x -> r^p * x / r, where r is the L2 norm of x.
}
\examples{
X <- matrix(runif(20), ncol = 2)
transformed_X <- radial.Lp.transform(X, p = 0.5)

}
