% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{radial.qexp.gaussian}
\alias{radial.qexp.gaussian}
\title{Evaluates radial q-exponential Gaussian on set of points}
\usage{
radial.qexp.gaussian(X, center, sigma = 1, q = 1)
}
\arguments{
\item{X}{A matrix of data frame of points at which the radia q-exponential Gaussian is to be evaluated.}

\item{center}{The center of the radial q-exponential Gaussian.}

\item{sigma}{The standard deviation of the radial q-exponential Gaussian.}

\item{q}{The power to which the radial q-exponential Gaussian is raised.}
}
\value{
Returns a vector of values of the radia q-exponential Gaussian at the points of X.
}
\description{
Evaluates radial q-exponential Gaussian cetnered at 'center' with standard deviation sigma
and power q on a set of points X.
}
