% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{remove.close.neighbors}
\alias{remove.close.neighbors}
\title{Removes Close Neighbors}
\usage{
remove.close.neighbors(x, min.dist)
}
\arguments{
\item{x}{A numeric vector.}

\item{min.dist}{A positive numeric value specifying the minimum allowable
distance between any two consecutive elements of x.}
}
\value{
A modified version of x such that no two consecutive elements are closer than \code{min.dist}.
}
\description{
This function takes a vector of positions (\code{x}) and a minimum distance
(\code{min.dist}), and returns a modified vector of knots where no two knots are
closer than the specified minimum distance.
}
\details{
The function first sorts the input vector of knots. It then iterates through
the sorted knots, keeping only those that are at least \code{min.dist} away from
the previously accepted knot. This ensures that the output vector of knots
satisfies the distance requirement.
}
\examples{
\dontrun{
x <- c(0.1, 0.2, 0.4, 0.7, 1.0)
min.dist <- 0.25
remove.close.neighbors(x, min.dist)
}
}
