% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_cx.R
\name{replace.basin.label.basin_cx}
\alias{replace.basin.label.basin_cx}
\title{Replace a Basin Label in a Basin Complex Object}
\usage{
replace.basin.label.basin_cx(bcx, old.label, new.label)
}
\arguments{
\item{bcx}{An object of class "basin_cx"}

\item{old.label}{The existing label to be replaced}

\item{new.label}{The new label to replace with}
}
\value{
A modified object of class "basin_cx" with updated labels
}
\description{
Replaces a specified label for a local extremum in a basin_cx object.
Updates all references to this label in extrema, basins, and cells.
}
