% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{restric.to.Linf.unit.sphere}
\alias{restric.to.Linf.unit.sphere}
\title{Restrict Points to L-infinity Unit Sphere}
\usage{
restric.to.Linf.unit.sphere(X)
}
\arguments{
\item{X}{A matrix or data frame containing numeric values to be restricted.}
}
\value{
A matrix or data frame of the same dimensions as the input, with each row
restricted to the L-infinity unit sphere.
}
\description{
This function restricts the points represented by rows in the input matrix or data frame
to the L-infinity unit sphere. If the maximum absolute value of a row exceeds 1,
the row is scaled down to have a maximum absolute value of 1.
}
\examples{
X <- matrix(runif(20, -2, 2), ncol = 2)
restricted_X <- restric.to.Linf.unit.sphere(X)

}
