% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{right.winsorize}
\alias{right.winsorize}
\title{Right Winsorize a numeric vector}
\usage{
right.winsorize(y, p = 0.01)
}
\arguments{
\item{y}{A numeric vector to be right-winsorized}

\item{p}{A numeric value between 0 and 0.25 indicating the proportion of data to be replaced on the right tail (default is 0.01)}
}
\value{
A numeric vector with extreme values on the right tail replaced
}
\description{
This function applies right-winsorization to a numeric vector, replacing extreme values on the right tail.
}
\examples{
y <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
right.winsorize(y, p = 0.1)

}
