% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{rm.self.loops}
\alias{rm.self.loops}
\title{Removes self-loops from a graph}
\usage{
rm.self.loops(adj.list)
}
\arguments{
\item{adj.list}{A named list representing the adjacency list of the graph. Each element of the
list is a numeric vector containing the indices of the adjacent vertices for a
given vertex. The names of the list elements correspond to the vertex names.}
}
\value{
A named list representing the adjacency list of the graph with self-loops removed.
The format is the same as the input adjacency list, but without any self-loops.
}
\description{
This function takes an adjacency list representation of a graph and removes all self-loops
from it. A self-loop is an edge that connects a vertex to itself.
}
\examples{
graph <- list(
  "A" = c(1, 2),
  "B" = c(2, 3),
  "C" = c(1, 3),
  "D" = c(4)
)

graph.no.self.loops <- rm.self.loops(graph)
print(graph.no.self.loops)

}
