% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{robust.z.normalize}
\alias{robust.z.normalize}
\title{Perform Robust Z-Score Normalization}
\usage{
robust.z.normalize(x)
}
\arguments{
\item{x}{A numeric vector to be normalized}
}
\value{
A numeric vector of the same length as the input containing robust z-scores.
Missing values (NA) in the input will result in NA in the output at the same positions.
}
\description{
Normalizes data using robust statistics by subtracting the median and dividing by the Median
Absolute Deviation (MAD). This method is more resistant to outliers compared to traditional
z-score normalization that uses mean and standard deviation.
}
\details{
The function implements robust z-score normalization using the formula:
z = (x - median(x)) / MAD(x)

The MAD is scaled by a factor of 1.4826 to make it consistent with the standard deviation
when the data is normally distributed.
}
\note{
If the MAD is zero, the function will return a vector of zeros and issue a warning.
}
\examples{
# Basic usage
data <- c(1, 2, 3, 100, 4, 5, 6)  # Note the outlier
robust.z.normalize(data)

# Handling missing values
data.with.na <- c(1, NA, 3, 100, 4, NA, 6)
robust.z.normalize(data.with.na)

}
