% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sampling.R
\name{runif.simplex}
\alias{runif.simplex}
\title{Produces a random sample from the uniform distribution over the (K-1)-dimensional simplex.}
\usage{
runif.simplex(K)
}
\arguments{
\item{K}{The size of vector of non-negative numbers that sum up to 1.}
}
\value{
A numeric vector of non-negative numbers that sum to 1.
}
\description{
Produces a random sample from the uniform distribution over the (K-1)-dimensional simplex.
}
