% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sampling.R
\name{runif.sphere}
\alias{runif.sphere}
\title{Uniformly Sample Points from a Unit Sphere Surface}
\usage{
runif.sphere(n.samples, dim)
}
\arguments{
\item{n.samples}{Integer > 0 indicating the number of samples to generate.}

\item{dim}{Integer >= 0 indicating the intrinsic dimension of the sphere.
The sphere will be embedded in (dim+1)-dimensional space.
E.g., dim=2 for points on a regular 3D sphere surface.}
}
\value{
A matrix of dimension \code{n.samples x (dim+1)} where each row represents
a point uniformly sampled from the surface of the unit sphere. All points
have unit Euclidean norm.
}
\description{
Generates uniform random samples from the surface of a unit sphere in (dim+1)-dimensional
space using the Gaussian normalization method. For example, dim=2 generates points on
the surface of a 3D unit ball (2-sphere).
}
\details{
The function uses the fact that normalizing a vector of independent standard normal
variables results in a uniform distribution on the sphere surface.
}
\examples{
# Generate 10 points on a 2D circle (1-sphere)
points_circle <- runif.sphere(10, 1)

# Generate 100 points on a 3D sphere surface (2-sphere)
points_sphere <- runif.sphere(100, 2)

}
\references{
Muller, M. E. (1959). A note on a method for generating points uniformly on
n-dimensional spheres. Communications of the ACM, 2(4), 19-20.
}
