% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_factor_analysis.R
\name{save_results}
\alias{save_results}
\title{Save Analysis Results in Multiple Formats}
\usage{
save_results(all.f, f, p, pval, out.dir, label, latex_file)
}
\arguments{
\item{all.f}{formatted results matrix}

\item{f}{frequency table}

\item{p}{proportion table}

\item{pval}{p-values vector}

\item{out.dir}{output directory}

\item{label}{file label}

\item{latex_file}{LaTeX output file path}
}
\description{
Internal function to save analysis results as RDA, CSV, and optionally LaTeX files.
}
\keyword{internal}
