% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{scale_to_range}
\alias{scale_to_range}
\title{Linear transformation of 'x' values so that the min(x) is sent to ymin and max(x) to ymax}
\usage{
scale_to_range(x, ymin, ymax, xrange = NULL)
}
\arguments{
\item{x}{Input values}

\item{ymin}{The value to which min(x) is sent.}

\item{ymax}{The value to which max(x) is sent.}

\item{xrange}{The minimum of x and the maximum of x that is to be mapped to ymin and ymax, respectively.}
}
\value{
A numeric vector of the same length as 'x' with values linearly transformed to the range \code{[ymin, ymax]}.
}
\description{
Linear transformation of 'x' values so that the min(x) is sent to ymin and max(x) to ymax
}
