% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{scaled.log.tan}
\alias{scaled.log.tan}
\title{Applies a Scaled Log Transformation}
\usage{
scaled.log.tan(x, scale = 0.1)
}
\arguments{
\item{x}{The input variable to undergo transformation.}

\item{scale}{A value representing the desired offset of the transformed variable range from -pi/2 and pi/2.}
}
\value{
Returns the tangent of the scaled log of x.
}
\description{
This function first applies a scaled log transformation, C*log(x), to the
input variable 'x'. The scaling is performed such that the range of the
transformed variable 'x' lies within the interval \code{[-pi/2, pi/2]}.
Subsequently, a tangent function is applied to the result of the
transformation.
}
\examples{
\dontrun{
x <- 0.1
scaled.log.tan(x)
}
}
