% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_smooth_morse_smale.R
\name{separatrices.plot}
\alias{separatrices.plot}
\title{Plot Separatrices}
\usage{
separatrices.plot(
  separatrices,
  M1_pos,
  M2_pos,
  line.lwd = 2,
  line.lty = 2,
  line.col = "black",
  add = TRUE,
  ...
)
}
\arguments{
\item{separatrices}{List of separatrix trajectories}

\item{M1_pos}{Position of first maximum}

\item{M2_pos}{Position of second maximum}

\item{line.lwd}{Line width (default 2)}

\item{line.lty}{Line type (default 2, dashed)}

\item{line.col}{Line color (default "black")}

\item{add}{Logical, whether to add to existing plot}

\item{...}{Additional arguments passed to plot}
}
\value{
Invisible NULL
}
\description{
Plots separatrices emanating from saddle points.
}
