% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerve_cx.R
\name{set.complex.function.values}
\alias{set.complex.function.values}
\title{Set Function Values on Nerve Complex Vertices}
\usage{
set.complex.function.values(complex, values)
}
\arguments{
\item{complex}{A nerve complex object}

\item{values}{Vector of function values (one per vertex)}
}
\value{
The updated nerve complex object (invisibly)
}
\description{
This function sets function values at the vertices of the nerve complex.
}
