% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerve_cx.R
\name{set.complex.weight.scheme}
\alias{set.complex.weight.scheme}
\title{Set Weight Scheme for Nerve Complex}
\usage{
set.complex.weight.scheme(complex, weight.type, params = numeric(0))
}
\arguments{
\item{complex}{A nerve complex object}

\item{weight.type}{Type of weight scheme to use}

\item{params}{Parameters for the weight scheme (if needed)}
}
\value{
The updated nerve complex object (invisibly)
}
\description{
This function sets the weighting scheme for simplices in the nerve complex.
}
\details{
Available weight types:
\itemize{
\item "uniform": All simplices have weight 1.0
\item "inverse_distance": Weights are inversely proportional to average squared distance
\item "gaussian": Weights use Gaussian kernel based on distances (param: sigma)
\item "volume": Weights based on approximate simplex volume (param: alpha)
\item "gradient": Weights based on function gradient over simplex (param: gamma)
}
}
