% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{shifted.tan}
\alias{shifted.tan}
\title{Applies a Linear Transformation and Tangent Function}
\usage{
shifted.tan(x, offset = 0.1)
}
\arguments{
\item{x}{The input variable to undergo transformation.}

\item{offset}{A value representing the desired offset of the transformed variable range from -pi/2 and pi/2.}
}
\value{
Returns the tangent of the linearly transformed 'x'.
}
\description{
This function applies a linear transformation to the input variable 'x',
scaling it so that the range of the transformed 'x' falls within the interval
\code{[-pi/2 + offset, pi/2 - offset]}. The function then applies a tangent
function to the result of the transformation.
}
\examples{
\dontrun{
x <- 0.001
shifted.tan(x)
}
}
