% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{skewed.gaussian}
\alias{skewed.gaussian}
\title{Compute Value of a Single Skewed Gaussian}
\usage{
skewed.gaussian(x, mu, sigma, alpha)
}
\arguments{
\item{x}{Numeric. The point at which to evaluate the skewed Gaussian.}

\item{mu}{Numeric. The location parameter (equivalent to xi in the formula).}

\item{sigma}{Numeric. The scale parameter (equivalent to omega in the formula).}

\item{alpha}{Numeric. The shape parameter controlling the skewness.}
}
\value{
Numeric. The value of the skewed Gaussian distribution at point x.
}
\description{
The skewed Gaussian distribution is defined by the following formula:

f(x | xi, omega, alpha) = (2 / omega) * phi((x - xi) / omega) * Phi(alpha * ((x - xi) / omega))

where phi is the standard normal density and Phi is the standard normal
cumulative distribution function.
}
\details{
This function calculates the value of a skewed Gaussian distribution at a given point.

The parameters of the distribution are:
\itemize{
\item xi (equivalent to mu): location parameter
\item omega (equivalent to sigma): scale parameter
\item alpha: shape parameter controlling the skewness
}
}
\examples{
skewed.gaussian(0, mu = 0, sigma = 1, alpha = 2)

}
