% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesics.R
\name{smooth.PL.geodesic}
\alias{smooth.PL.geodesic}
\title{Smooths a piece-wise linear (PL) geodesic path}
\usage{
smooth.PL.geodesic(
  S,
  path.vs,
  min.vs = 10,
  adj.n.vs = 50,
  grid.size = 200,
  bw = grid.size * 3/200,
  fb.C = 3
)
}
\arguments{
\item{S}{A state space.}

\item{path.vs}{A vector of indices of S constituting an geodesic path.}

\item{min.vs}{The minimum number of vertices in path.vs for which a smoothing can be done. If |path.vs| < min.vs, then approx() is used to generate adj.n.vs vertices.}

\item{adj.n.vs}{The number of vertices that is being created as the input for fb.magelo() in case |path.vs| < min.vs.}

\item{grid.size}{The number of grid points; A parameter passed to magelo().}

\item{bw}{The bandwidth of fb.magelo().}

\item{fb.C}{The number of bw's away from the boundary points of the domain of x that the adjusting of Eyg will take place, so that Eyg satisfies the boundary condition.}
}
\value{
A matrix representing the smoothed geodesic path with grid.size rows and
the same number of columns as the state space S. Each row contains the coordinates
of a point along the smoothed path.
}
\description{
Smooths a piece-wise linear (PL) geodesic path
}
\details{
This function takes a piecewise linear geodesic path defined by vertex indices and
produces a smooth version using local linear regression. If the input path has fewer
than min.vs vertices, it first interpolates to create adj.n.vs points before smoothing.
The smoothing is performed separately for each dimension of the state space.
}
