% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesics.R
\name{subdivide.path}
\alias{subdivide.path}
\title{Subdivides a path of points in \eqn{R^n} into a uniform grid of points along
the path}
\usage{
subdivide.path(path, n.subdivision.pts)
}
\arguments{
\item{path}{A matrix of consecutive path points in some state space.}

\item{n.subdivision.pts}{The number of points in the grid.}
}
\value{
A matrix with n.subdivision.pts rows and the same number of columns as the
input path. Each row represents a point along the subdivided path, with points
evenly spaced according to the path's arc length.
}
\description{
Subdivides a path of points in \eqn{R^n} into a uniform grid of points along
the path
}
\details{
This function takes a path defined by consecutive points and creates a new set of
points that are uniformly distributed along the path's arc length. The subdivision
is based on the cumulative Euclidean distances between consecutive points, ensuring
that the output points are evenly spaced along the actual path length rather than
just parametrically.

The first point of the output will be the first point of the input path, and the
last point of the output will be the last point of the input path.
}
\examples{
# Create a simple 2D path
path <- matrix(c(0,0, 1,0, 2,1, 3,1), ncol=2, byrow=TRUE)
# Subdivide into 10 evenly spaced points
subdivided <- subdivide.path(path, 10)

}
