% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesics.R
\name{subdivide.path.v1}
\alias{subdivide.path.v1}
\title{Subdivides a path of points in \eqn{R^n} into a uniform grid of points along the path}
\usage{
subdivide.path.v1(path, n.subdivision.pts)
}
\arguments{
\item{path}{A matrix of consecutive path points in some state space.}

\item{n.subdivision.pts}{The number of points in the grid.}
}
\value{
A matrix with n.subdivision.pts rows and the same number of columns as the
input path. Each row represents a point along the subdivided path, with points
evenly spaced according to the path's arc length.
}
\description{
Subdivides a path of points in \eqn{R^n} into a uniform grid of points along the path
}
\details{
This is version 1 of the path subdivision algorithm. Like subdivide.path, it creates
a set of points uniformly distributed along the path's arc length. The implementation
differs slightly in the algorithm used but produces similar results.
}
\note{
This appears to be an alternative implementation of subdivide.path. Consider
using subdivide.path unless there's a specific reason to use this version.
}
\seealso{
\code{\link{subdivide.path}} for the primary implementation
}
