% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic_stats.R
\name{summarize_overlap_statistics}
\alias{summarize_overlap_statistics}
\title{Summarize overlap statistics from geodesic_stats object}
\usage{
summarize_overlap_statistics(object)
}
\arguments{
\item{object}{A geodesic_stats object containing overlap statistics}
}
\value{
A data frame with summary statistics for overlap at each radius
}
\description{
Summarize overlap statistics from geodesic_stats object
}
\keyword{internal}
