% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fassoc0_test.R
\name{summary.assoc0}
\alias{summary.assoc0}
\title{Summary Method for assoc0 Objects}
\usage{
\method{summary}{assoc0}(object, ...)
}
\arguments{
\item{object}{An object of class "assoc0".}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A list of class "summary.assoc0" with summary statistics.
}
\description{
Summary Method for assoc0 Objects
}
