% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fassoc1_test.R
\name{summary.assoc1}
\alias{summary.assoc1}
\title{Summary Method for assoc1 Objects}
\usage{
\method{summary}{assoc1}(object, ...)
}
\arguments{
\item{object}{An object of class "assoc1".}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A list of class "summary.assoc1" with summary statistics.
}
\description{
Summary Method for assoc1 Objects
}
