% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fassoc_test.R
\name{summary.fassoc}
\alias{summary.fassoc}
\title{Summary Method for Functional Association Tests}
\usage{
\method{summary}{fassoc}(object, ...)
}
\arguments{
\item{object}{An object of class "assoc0" or "assoc1".}

\item{...}{Additional arguments passed to specific summary methods.}
}
\value{
A summary object.
}
\description{
Summary Method for Functional Association Tests
}
