% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gflow_basins.R
\name{summary.gflow_basins}
\alias{summary.gflow_basins}
\title{Summarize Gradient-Flow Basin Results}
\usage{
\method{summary}{gflow_basins}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"gflow_basins"}}

\item{...}{Additional arguments (currently ignored)}
}
\value{
An object of class \code{"summary.gflow_basins"} containing:
\describe{
\item{n_extrema}{Total number of extrema detected}
\item{n_minima}{Number of local minima}
\item{n_maxima}{Number of local maxima}
\item{n_ascending_basins}{Number of ascending basins}
\item{n_descending_basins}{Number of descending basins}
\item{ascending_basin_sizes}{Summary statistics for ascending basin sizes}
\item{descending_basin_sizes}{Summary statistics for descending basin sizes}
}
}
\description{
Summarize Gradient-Flow Basin Results
}
