% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spectral_filter.R
\name{summary.graph_spectral_filter}
\alias{summary.graph_spectral_filter}
\title{Summary Method for Graph Spectral Filter Results}
\usage{
\method{summary}{graph_spectral_filter}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"graph_spectral_filter"}}

\item{...}{Further arguments passed to or from other methods}
}
\value{
An object of class \code{"summary.graph_spectral_filter"} containing
key summary statistics
}
\description{
Summary Method for Graph Spectral Filter Results
}
