% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detection.R
\name{summary.knn.outliers}
\alias{summary.knn.outliers}
\title{Summary Method for knn.outliers Objects}
\usage{
\method{summary}{knn.outliers}(object, ...)
}
\arguments{
\item{object}{An object of class "knn.outliers" as returned by \code{remove.knn.outliers}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A list of class "summary.knn.outliers" containing summary statistics.
}
\description{
Provides a detailed summary of the outlier detection results from \code{remove.knn.outliers}.
}
\examples{
# Using the example from remove.knn.outliers
set.seed(123)
S <- rbind(matrix(rnorm(2000), ncol = 2),
           cbind(rnorm(10, 10), rnorm(10, 10)))
result <- remove.knn.outliers(S, K = 10)
summary(result)

}
