% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uggmalo.R
\name{summary.uggmalo}
\alias{summary.uggmalo}
\title{Summary Method for UGGMALO Results}
\usage{
\method{summary}{uggmalo}(object, ...)
}
\arguments{
\item{object}{An object of class "uggmalo"}

\item{...}{Additional arguments (currently ignored)}
}
\value{
A list of class "summary.uggmalo" containing summary statistics
}
\description{
Summary Method for UGGMALO Results
}
