% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uggmalog.R
\name{summary.uggmalog}
\alias{summary.uggmalog}
\title{Summary method for uggmalog objects}
\usage{
\method{summary}{uggmalog}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"uggmalog"}}

\item{...}{Additional arguments (currently ignored)}
}
\value{
A list with summary information, invisibly
}
\description{
Summary method for uggmalog objects
}
