% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic_stats.R
\name{summary.vertex_geodesic_stats}
\alias{summary.vertex_geodesic_stats}
\title{Summary method for vertex_geodesic_stats objects}
\usage{
\method{summary}{vertex_geodesic_stats}(object, ...)
}
\arguments{
\item{object}{A vertex_geodesic_stats object}

\item{...}{Additional arguments passed to summary}
}
\value{
A summary of the vertex geodesic statistics
}
\description{
Summary method for vertex_geodesic_stats objects
}
