% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{synthetic.mixture.of.gaussians}
\alias{synthetic.mixture.of.gaussians}
\title{Generate a Synthetic Smooth Function in One Dimension}
\usage{
synthetic.mixture.of.gaussians(x, x.knot, y.knot = NULL, sd.knot = NULL)
}
\arguments{
\item{x}{Values at which the function is to be evaluated.}

\item{x.knot}{A numeric vector of the centers of the Gaussian functions.}

\item{y.knot}{An optional numeric vector of response values at the locations specified in \code{x.knot}.
If \code{y.knot} is not provided, random values will be generated.}

\item{sd.knot}{Standard deviations of the Gaussians.}
}
\value{
A function representing the synthetic smooth function. This function takes a numeric
value as input and returns a numeric value, representing the function's output at that point.
}
\description{
This function constructs a synthetic smooth function using a series of Gaussian distributions
in one-dimensional space. It is centered at given points with standard deviations determined
by the distances to neighboring points. The function can handle both positive and negative
Gaussian distributions.
}
\details{
For each point in \code{x}, the function calculates the standard
deviation of the Gaussian distribution as half of the minimum distance to
its neighbors. It then forms a synthetic function by combining these
Gaussian distributions. The function can generate smooth functions that
take also negative values by multiplying Gaussian components by a vector
of \eqn{\pm 1} values.
}
\examples{
\dontrun{
# Example usage
x <- seq(0, 1, length.out = 10)
synthetic.fn <- generate.synthetic.function(x)
point <- 0.5 # A point in 1D space
value <- synthetic.fn(point) # Evaluate the function at the given point
}
}
