% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{test.graph.MS.cx.on.synth.cloud.data}
\alias{test.graph.MS.cx.on.synth.cloud.data}
\title{Tests graph Morse-Smale complex construct (its persistence and reconstruction) on synthetic datasets}
\usage{
test.graph.MS.cx.on.synth.cloud.data(
  n.datasets,
  n.pts,
  n.side = NULL,
  dim = 2,
  type = "grid",
  Ks,
  n.lmax = 10,
  y.min = 1,
  y.max = 15,
  out.dir,
  verbose = TRUE
)
}
\arguments{
\item{n.datasets}{The number of datasets to generate.}

\item{n.pts}{The number of points the synthetic cloud dataset suppose to have.}

\item{n.side}{If the data is to be a sample from some grid, then n.side is the number of points on the axis of each dimension, thus the total number of grid points is n.size^dim.}

\item{dim}{The dimension of R^dim in which data is constructed.}

\item{type}{The method used for the construction.}

\item{Ks}{A vector of positive integer values specifying the range of k values.}

\item{n.lmax}{The number of gaussians to use.}

\item{y.min}{The mimimum value of the range of y values.}

\item{y.max}{The maximum value of the range of y values.}

\item{out.dir}{The output directory where the generated data is stored.}

\item{verbose}{Set it to TRUE, to print progress messages.}
}
\value{
A list containing the length of the longest stretch of isomorphic MS graphs for each generated dataset
}
\description{
Tests graph Morse-Smale complex construct (its persistence and reconstruction) on synthetic datasets
}
