% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spectral_lowess.R
\name{update.graph.spectral.lowess}
\alias{update.graph.spectral.lowess}
\title{Update Graph Spectral LOWESS Fit}
\usage{
\method{update}{graph.spectral.lowess}(object, formula, ..., evaluate = TRUE)
}
\arguments{
\item{object}{A 'graph.spectral.lowess' object}

\item{formula}{Changes to the model (currently unused)}

\item{...}{Additional arguments to override in the new call}

\item{evaluate}{Should the updated call be evaluated? (default: TRUE)}
}
\value{
Updated graph.spectral.lowess object (if evaluate = TRUE) or updated call
}
\description{
Updates a graph spectral LOWESS fit with new parameters
}
