% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{v1.torus.knot}
\alias{v1.torus.knot}
\title{Torus knot}
\usage{
v1.torus.knot(n, k, m)
}
\arguments{
\item{n}{The desired number of points to generate.}

\item{k}{Torus knot first parameter.}

\item{m}{Torus knot second parameter.}
}
\description{
Torus knot
}
