% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesics.R
\name{vars.approx.monotonically.assoc.with.geodesic}
\alias{vars.approx.monotonically.assoc.with.geodesic}
\title{Given TA and TAA from E.geodesic.X(), this routine identifies variables monotonically associated with the distance along the geodesic}
\usage{
vars.approx.monotonically.assoc.with.geodesic(X, eps = 0.4)
}
\arguments{
\item{X}{A matrix of a data frame with TA and delta1 columns.}

\item{eps}{The positive constant specifying the size of a neighborhood around the diagonal and anti-diagonal of the (TA, TAA) graph.}
}
\value{
A logical vector of length equal to the number of rows in X, where TRUE
indicates that the corresponding variable is approximately monotonically associated
with the geodesic distance (either increasing or decreasing).
}
\description{
Given TA and TAA from E.geodesic.X(), this routine identifies variables monotonically associated with the distance along the geodesic
}
\details{
This function analyzes the relationship between variables and geodesic distance by
examining the TA (turn angle) and TAA (turn angle acceleration) values. Variables
are considered monotonically associated if their (TA, TAA) values fall within an
epsilon neighborhood of either the diagonal (monotonically increasing) or
anti-diagonal (monotonically decreasing) in the unit square.
}
\examples{
# Assuming X has columns TA and TAA from E.geodesic.X output
# monotonic_vars <- vars.approx.monotonically.assoc.with.geodesic(X, eps = 0.4)
# X[monotonic_vars, ] # subset to monotonic variables

}
