% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{vert.error.bar}
\alias{vert.error.bar}
\title{Add Vertical Error Bar to Plot}
\usage{
vert.error.bar(x, ymin, ymax, dx = 0.025, lwd = 1, col = "red")
}
\arguments{
\item{x}{X-coordinate for the error bar.}

\item{ymin}{Lower limit of error bar.}

\item{ymax}{Upper limit of error bar.}

\item{dx}{Horizontal offset for end caps.}

\item{lwd}{Line width.}

\item{col}{Color of error bar.}
}
\value{
Invisibly returns NULL.
}
\description{
Adds a vertical error bar to an existing 2D plot
}
\examples{
plot(1:10, rnorm(10), ylim = c(-3, 3))
vert.error.bar(5, -1, 1, col = "red")

}
