% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein_dist.R
\name{wasserstein.ipNNuv}
\alias{wasserstein.ipNNuv}
\title{Local Wasserstein distance to the uniform and delta at 1 of the inner products of the NN unit vectors}
\usage{
wasserstein.ipNNuv(
  S,
  K = 25,
  use.transport = TRUE,
  n.breaks = 50,
  ref.i = 2813,
  use.geodesic.knn = TRUE,
  nn.i = NULL,
  nn.d = NULL,
  verbose = TRUE
)
}
\arguments{
\item{S}{A state space.}

\item{K}{The number of nearest neighbors to use for the identification of the end points}

\item{use.transport}{Set to TRUE, to use wasserstein1d() from transport package.}

\item{n.breaks}{The number of bin for the estimate of distribution of cosine of the angle values.}

\item{ref.i}{The index of the reference point whose cosine angles will be used as the reference for the Wd1 estimates.}

\item{use.geodesic.knn}{Set to TRUE, if geodesic.knn() is to be used instead of get.knn().}

\item{nn.i}{A matrix of indices of the nearest neighbors.}

\item{nn.d}{A matrix of distances to the nearest neighbors.}

\item{verbose}{Set to TRUE for messages indicating which part of the routine is currently run.}
}
\value{
A list with cosine of the angle values for each state of the state space.
}
\description{
For each point x of a state space, S, the unit vectors starting at x and
ending at x's K-NN's are defined. The routine estimates the entropy of the
inner products <n_i, n_K> of all these unit vectors with the last one.
}
