% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_utils.R
\name{plot3d.gaussian_mixture}
\alias{plot3d.gaussian_mixture}
\title{Create 3D interactive plot of gaussian_mixture}
\usage{
\method{plot3d}{gaussian_mixture}(x, n.grid = 50, col = "skyblue", alpha = 0.7, add.contours = TRUE, ...)
}
\arguments{
\item{x}{A gaussian_mixture object}

\item{n.grid}{Number of grid points in each dimension}

\item{col}{Surface color}

\item{alpha}{Surface transparency}

\item{add.contours}{Whether to add contour lines at the base}

\item{...}{Additional arguments passed to rgl functions}
}
\value{
Invisible surface object
}
\description{
Create 3D interactive plot of gaussian_mixture
}
